% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R
\name{as.list}
\alias{as.list}
\title{as.list method for DTAColumnSpec}
\usage{
## S7 method for class <DTAtools::DTAColumnSpec>
as.list(x, ...)
}
\arguments{
\item{x}{A DTAColumnSpec object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A named list with the DTAColumnSpec properties.
}
\description{
Converts a DTAColumnSpec object to a named list.
}
