% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecStructureSAS-class.R
\name{as.list}
\alias{as.list}
\title{as.list method for as.list.DTAColumnSpecStructureSAS}
\usage{
## S7 method for class <DTAtools::DTAColumnSpecStructureSAS>
as.list(x, ...)
}
\arguments{
\item{x}{A DTAColumnSpecStructureSAS object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A named list with the DTAColumnSpecStructureSAS properties.
}
\description{
Converts a DTAColumnSpecStructureSAS object to a named list.
}
