% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARule-class.R
\name{as.list}
\alias{as.list}
\title{as.list for DTARule}
\arguments{
\item{x}{An object of class DTARule.}

\item{...}{Additional arguments (not used).}
}
\value{
A named list containing the properties of the DTARule object.
}
\description{
Convert a DTARule object to a list.
}
