% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColRange-class.R
\name{as.list}
\alias{as.list}
\title{as.list for DTARuleColRange}
\usage{
## S7 method for class <DTAtools::DTARuleColRange>
as.list(x)
}
\arguments{
\item{x}{An object of class DTARuleColRange}

\item{...}{Additional arguments (not used).}
}
\value{
A named list containing the properties of the DTARuleColRange object.
}
\description{
Convert a DTARuleColRange object to a list.
}
