% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R
\name{as.list.DTAColumnSpec}
\alias{as.list.DTAColumnSpec}
\title{as.list method for DTAColumnSpec}
\usage{
\method{as.list}{DTAColumnSpec}(x, ...)
}
\arguments{
\item{x}{A DTAColumnSpec object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A named list with the DTAColumnSpec properties.
}
\description{
Converts a DTAColumnSpec object to a named list.
}
