% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecStructure-class.R
\name{as.list.DTAColumnSpecStructure}
\alias{as.list.DTAColumnSpecStructure}
\title{as.list method for as.list.DTAColumnSpecStructure}
\usage{
\method{as.list}{DTAColumnSpecStructure}(x, ...)
}
\arguments{
\item{x}{A as.list.DTAColumnSpecStructure object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A named list with the DTAColumnSpecStructure properties.
}
\description{
Converts a as.list.DTAColumnSpecStructure object to a named list.
}
