% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColCondition-class.R,
%   R/DTARuleColRange-class.R, R/DTARuleColUnique-class.R
\name{as.list.DTARuleColCondition}
\alias{as.list.DTARuleColCondition}
\title{as.list for DTARuleColCondition}
\usage{
\method{as.list}{DTARuleColCondition}(x, ...)

\method{as.list}{DTARuleColCondition}(x, ...)

\method{as.list}{DTARuleColCondition}(x, ...)
}
\arguments{
\item{x}{An object of class DTARuleColCondition}

\item{...}{Additional arguments (not used).}
}
\value{
A named list containing the properties of the DTARuleColCondition object.

A named list containing the properties of the DTARuleColCondition object.

A named list containing the properties of the DTARuleColCondition object.
}
\description{
Convert a DTARuleColCondition object to a list.

Convert a DTARuleColCondition object to a list.

Convert a DTARuleColCondition object to a list.
}
