% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleCollection-class.R
\name{as.list.DTARuleCollection}
\alias{as.list.DTARuleCollection}
\title{as.list for DTARuleCollection}
\usage{
\method{as.list}{DTARuleCollection}(x, ...)
}
\arguments{
\item{x}{An object of class DTARuleCollection}

\item{...}{Additional arguments (not used).}
}
\value{
A named list containing the properties of the DTARuleCollection object.
}
\description{
Convert a DTARuleCollection object to a list.
}
