% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFunctions.R
\name{change_type}
\alias{change_type}
\title{Change Type}
\usage{
change_type(vector, var_type)
}
\arguments{
\item{vector}{A vector of values.}

\item{var_type}{Character. The desired type of the vector.}
}
\value{
The vector converted to the specified type.
}
\description{
This function changes the type of a vector to the specified type.
}
\examples{
change_type(c("1", "2", "3"), "numeric")
}
