% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFunctions.R
\name{changeNAs}
\alias{changeNAs}
\title{Change NAs}
\usage{
changeNAs(vector, var_type)
}
\arguments{
\item{vector}{A vector of values.}

\item{var_type}{Character. The type of the vector.}
}
\value{
The modified vector with NAs replaced by an empty string if the type is character.
}
\description{
This function replaces NA values in a vector with an empty string if the vector type is character.
}
\examples{
\dontrun{
changeNAs(c(NA, "b", "c"), "character")
}
}
