% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetFile-class.R
\name{check,DTAtools::DTADataSetFile-method}
\alias{check,DTAtools::DTADataSetFile-method}
\title{Check DTADataSetFile}
\usage{
## S7 method for class <DTAtools::DTADataSetFile>
check(
  x,
  tables = NULL,
  force = FALSE,
  persist = TRUE,
  artifact_dir = NULL,
  quiet = FALSE,
  validation_run = NULL
)
}
\description{
Check DTADataSetFile
}
