% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFunctions.R
\name{checkFormat}
\alias{checkFormat}
\title{Check Format}
\usage{
checkFormat(value, var_format)
}
\arguments{
\item{value}{A vector of values.}

\item{var_format}{Integer. The maximum length allowed for each element.}
}
\value{
Logical. TRUE if all elements are within the specified format length, FALSE otherwise.
}
\description{
This function checks if the length of each element in a vector is less than or equal to a specified format length.
}
\examples{
checkFormat(c("abc", "defg"), 4)
}
