% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFunctions.R
\name{checkNullable}
\alias{checkNullable}
\title{Check Nullable}
\usage{
checkNullable(vector, nullable)
}
\arguments{
\item{vector}{A vector of values.}

\item{nullable}{Logical. Indicates if the vector can contain null values.}
}
\value{
Logical. TRUE if the vector meets the nullable condition, FALSE otherwise.
}
\description{
This function checks if a vector contains non-null values when nullable is FALSE.
}
\examples{
checkNullable(c(1, 2, NA), FALSE)
}
