% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFunctions.R
\name{checkType}
\alias{checkType}
\title{Check Type}
\usage{
checkType(vector, var_type)
}
\arguments{
\item{vector}{A vector of values.}

\item{var_type}{Character. The expected type of the vector.}
}
\value{
Logical. TRUE if the vector matches the specified type, FALSE otherwise.
}
\description{
This function checks if the type of a vector matches the specified type.
}
\examples{
checkType(c(1, 2, 3), "numeric")
}
