% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{colspec-DTADataSetTabular}
\alias{colspec-DTADataSetTabular}
\alias{colspec}
\title{Get Column by ID Method}
\usage{
colspec(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}

\item{id}{Character. The ID of the column to retrieve.}
}
\value{
A DTAColumnSpec object corresponding to the specified ID.
}
\description{
Method to get a column format by its ID from the collection.
}
\examples{
\dontrun{
column_format <- column(dtadata, "STUDYID")
}
}
