% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R,
%   R/DTADataSetTabular-class.R
\name{colspec}
\alias{colspec}
\title{Get Column by ID Method}
\usage{
colspec(x, ...)

colspec(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}

\item{...}{Additional named arguments:
\describe{
\item{id}{Character. The ID of the column to retrieve.}
}}
}
\value{
A DTAColumnSpec object corresponding to the specified ID.

A DTAColumnSpec object corresponding to the specified ID.
}
\description{
Method to get a column format by its ID from the collection.

Method to get a column format by its ID from the collection.
}
\examples{
collection <- create_example_DTAColumnSpecCollection()
colspec(collection, "STUDYID")
ds <- create_example_DTADataSetTabular(2)
colspec(ds, "STUDYID")
}
