% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{columns}
\alias{columns}
\title{Get column specs from DTADataSetTabular Object}
\usage{
columns(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}
}
\value{
A list with metadata information
}
\description{
Method to get columns specifications from DTADataSetTabular
}
\examples{
library(DTAtools)
ds <- create_example_DTADataSetTabular()
columns(ds)
}
