% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{columns}
\alias{columns}
\title{Get column specs from DTADataSetTabular Object}
\usage{
columns(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}

\item{...}{Not used by current methods; reserved for future extensions.}
}
\value{
A named list of \code{\link{DTAColumnSpec}} objects, one per
declared column, named by column id.
}
\description{
Method to get columns specifications from DTADataSetTabular
}
\examples{
ds <- create_example_DTADataSetTabular()
columns(ds)
}
