% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{columns_specs_from_word}
\alias{columns_specs_from_word}
\title{Create DTAColumnSpecCollection from DTA Word Document}
\usage{
columns_specs_from_word(
  file,
  table_position = 1,
  colnames = c("id", "label", "type", "length", "format", "nullable", "description"),
  value_sep = ";"
)
}
\arguments{
\item{file}{Character. Path to the Word document.}

\item{table_position}{Integer. Index of the table to extract.}

\item{colnames}{Vector. Vector containing column names of the table. Essential column names are: id Variable Name), label (Variable Label), type (Type), nullable (Nullable), description (Description)}

\item{value_sep}{Character. Separator dividing the values. Default: ";"}
}
\value{
A DTAColumnSpecCollection object.
}
\description{
Parses a DTA Word document to extract column specifications and create a DTAColumnSpecCollection object.
}
\examples{
\dontrun{
  import_columns_from_word(file)
}
}
