% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{columns_to_jsonschema}
\alias{columns_to_jsonschema}
\title{Convert DTAColumnSpec s to JSON Schema}
\usage{
columns_to_jsonschema(columns)
}
\arguments{
\item{columns}{Column spec information}
}
\value{
A list representing the JSON Schema.
}
\description{
Converts a DTAColumnSpec s into a JSON Schema.
}
