% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{container}
\alias{container}
\title{Get container}
\arguments{
\item{x}{An object of class DTA.}

\item{name}{Optional single character or single integer. if NULL, returns a
list of all containers. If character, returns the container with the specified name.
If integer, returns the container at the specified index.}
}
\value{
Either a list of DTADataSet objects or a single DTADataSet.
}
\description{
Method to get one or more containers from a DTA object.
}
\examples{
libary(DTAtools)
x <- create_example_DTA()
container(x)
container(x, "vitals")
container(x, 1)
}
