% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARule-class.R
\name{create_DTARule}
\alias{create_DTARule}
\title{Create a DTARule Object}
\usage{
create_DTARule(id, type, ...)
}
\arguments{
\item{id}{A character string specifying the rule identifier.}

\item{type}{A character string specifying the type of rule to create.
Supported types are \code{"col_condition"}, \code{"col_range"}, and \code{"col_unique"}.}

\item{...}{Additional arguments passed to the specific DTARule constructor.}
}
\value{
An object of class \code{DTARuleColCondition}, \code{DTARuleColRange}, or \code{DTARuleColUnique}, depending on \code{type}.
}
\description{
Factory function to create a DTARule object of a specified type.
}
\examples{
create_DTARule("rule1", "col_condition", column = "age", condition = "18", then = list("status" = list("equals" = "21"))) #TODO check
create_DTARule("rule2", "col_range", column = "score", min = 0, max = 100)
create_DTARule("rule3", "col_unique", column = "id")
}
