% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{create_example_DTA}
\alias{create_example_DTA}
\title{Create Example DTA Object}
\usage{
create_example_DTA(index = 1)
}
\arguments{
\item{index}{index of the example to create}
}
\value{
An object of class DTA with example data
}
\description{
Creates an example DTA object for demonstration purposes.
}
\examples{
example_dta <- create_example_DTA()
print(example_dta)
}
