% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R
\name{create_example_DTAColumnSpec}
\alias{create_example_DTAColumnSpec}
\title{Create Example DTAColumnSpec}
\usage{
create_example_DTAColumnSpec(index = 1)
}
\arguments{
\item{index}{Numeric. Selector for different example objects.}
}
\value{
An example DTAColumnSpec object based on the provided index.
}
\description{
S7 method to create and return an example DTAColumnSpec object.
}
\examples{
create_example_DTAColumnSpec()
}
