% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{create_example_DTAColumnSpecCollection}
\alias{create_example_DTAColumnSpecCollection}
\title{Create Example DTAColumnSpecCollection}
\usage{
create_example_DTAColumnSpecCollection(index = 1)
}
\arguments{
\item{index}{Integer. Index of the example to create.}
}
\value{
An example DTAColumnSpecCollection object.
}
\description{
S7 method to create and return an example DTAColumnSpecCollection object.
}
\examples{
create_example_DTAColumnSpecCollection()
}
