% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{create_example_DTADataSet}
\alias{create_example_DTADataSet}
\title{Create Example DTADataSet}
\usage{
create_example_DTADataSet(index = 1)
}
\arguments{
\item{index}{Integer. Index of the example to create.}
}
\value{
An example DTADataSet object.
}
\description{
S7 method to create and return an example DTADataSet object.
}
\examples{
library(DTAtools)
create_example_DTADataSet()
}
