% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfoCSV-class.R
\name{create_example_DTAFileInfoCSV}
\alias{create_example_DTAFileInfoCSV}
\title{Create Example DTAFileInfoCSV Object}
\usage{
create_example_DTAFileInfoCSV(index = 1)
}
\arguments{
\item{index}{example selector.}
}
\value{
An example \code{DTAFileInfoCSV} object.
}
\description{
Creates an example \code{DTAFileInfoCSV} object using example files.
}
\examples{
library(DTAtools)
create_example_DTAFileInfoCSV()
}
