% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfoTSV-class.R
\name{create_example_DTAFileInfoTSV}
\alias{create_example_DTAFileInfoTSV}
\title{Create Example DTAFileInfoTSV Object}
\usage{
create_example_DTAFileInfoTSV(index = 1)
}
\arguments{
\item{index}{example selector.}
}
\value{
An example \code{DTAFileInfoTSV} object.
}
\description{
Creates an example \code{DTAFileInfoTSV} object using example files.
}
\examples{
library(DTAtools)
create_example_DTAFileInfoTSV()
}
