% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileTSV-class.R
\name{create_example_DTAFileTSV}
\alias{create_example_DTAFileTSV}
\title{Create Example DTAFileTSV Object}
\usage{
create_example_DTAFileTSV(index = 1)
}
\arguments{
\item{index}{example selector.}
}
\value{
An example \code{DTAFileTSV} object.
}
\description{
Creates an example \code{DTAFileTSV} object using example files.
}
\examples{
create_example_DTAFileTSV()
}
