% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleCheckColCondition-class.R
\name{create_example_DTARuleCheckColCondition}
\alias{create_example_DTARuleCheckColCondition}
\title{create_example_DTARuleCheckColCondition}
\usage{
create_example_DTARuleCheckColCondition(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARule
}
\examples{
 library(DTAtools)
 create_example_DTARuleCheckColCondition()
}
