% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleCheckRange-class.R
\name{create_example_DTARuleCheckRange}
\alias{create_example_DTARuleCheckRange}
\title{create_example_DTARuleCheckRange}
\usage{
create_example_DTARuleCheckRange(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARuleCheckRange
}
\examples{
 library(DTAtools)
 create_example_DTARuleCheckRange()
}
