% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleCheckUnique-class.R
\name{create_example_DTARuleCheckUnique}
\alias{create_example_DTARuleCheckUnique}
\title{create_example_DTARuleCheckUnique}
\usage{
create_example_DTARuleCheckUnique(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARuleCheckUnique
}
\examples{
 library(DTAtools)
 create_example_DTARuleCheckUnique()
}
