% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleCollection-class.R
\name{create_example_DTARuleCollection}
\alias{create_example_DTARuleCollection}
\title{create_example_DTARuleCollection}
\usage{
create_example_DTARuleCollection(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARuleCollection
}
\examples{
 library(DTAtools)
 create_example_DTARuleCollection()
}
