% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColRange-class.R
\name{create_example_DTARuleColRange}
\alias{create_example_DTARuleColRange}
\title{Create an Example Column-Range Rule}
\usage{
create_example_DTARuleColRange(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\value{
An example \code{DTARuleColRange} object.
}
\description{
create example for DTARuleColRange
}
\examples{
create_example_DTARuleColRange()
}
