% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColUnique-class.R
\name{create_example_DTARuleColUnique}
\alias{create_example_DTARuleColUnique}
\title{Create an Example Column-Uniqueness Rule}
\usage{
create_example_DTARuleColUnique(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\value{
An example \code{DTARuleColUnique} object.
}
\description{
create example for DTARuleColUnique
}
\examples{
create_example_DTARuleColUnique()
}
