% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARule-class.R
\name{create_example_rule}
\alias{create_example_rule}
\title{create_example_rule}
\usage{
create_example_rule(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARule
}
\examples{
 library(DTAtools)
 create_example_rule()
}
