% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{datasets}
\alias{datasets}
\title{Get datasets}
\usage{
datasets(x, ...)
}
\arguments{
\item{x}{An object of class DTA.}

\item{...}{Additional named arguments:
\describe{
\item{name}{Optional single character or single integer. If NULL
(the default), returns a list of all datasets. If character, returns
the dataset with that name. If integer, returns the dataset at that
index.}
}}
}
\value{
Either a list of DTADataSet objects or a single DTADataSet.
}
\description{
Method to get one or more datasets from a DTA object.
}
\examples{
x <- create_example_DTA()
datasets(x)
datasets(x, "vitals")
datasets(x, 1)
}
