% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{datatable-DTAContainer}
\alias{datatable-DTAContainer}
\alias{datatable}
\title{Get table from DTAContainer Object}
\usage{
datatable(x, id = 1, ...)
}
\arguments{
\item{x}{An object of class DTAContainer.}

\item{id}{Character or numeric. Name or index of the table to retrieve.}
}
\value{
A Table object
}
\description{
Extract a table from the tables in a DTAContainer object.
}
\examples{
\dontrun{
datatable(container)           # returns first table
datatable(container, "lab")   # returns table named "lab"
}
}
