% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentBuilders.R
\name{.add_authorized_for_corrections_section}
\alias{.add_authorized_for_corrections_section}
\title{Add "Authorized for Corrections" section}
\usage{
.add_authorized_for_corrections_section(doc, authorized_for_corrections)
}
\description{
Add "Authorized for Corrections" section
}
\keyword{internal}
