% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattingHelpers.R
\name{.format_condition_map}
\alias{.format_condition_map}
\title{Render a whole condition map (column -> constraint) as one clause.}
\usage{
.format_condition_map(map)
}
\description{
Render a whole condition map (column -> constraint) as one clause.
}
\keyword{internal}
