% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.normalize_template_variables}
\alias{.normalize_template_variables}
\title{Normalize a user-supplied variable list to brace-delimited character values}
\usage{
.normalize_template_variables(variables, dta = NULL)
}
\description{
Normalize a user-supplied variable list to brace-delimited character values
}
\keyword{internal}
