% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-class.R
\name{.process_transmission_dates}
\alias{.process_transmission_dates}
\title{Process Transmission Dates}
\usage{
.process_transmission_dates(transmission)
}
\arguments{
\item{transmission}{A transmission list.}
}
\value{
A list with \code{transmission} (the processed list) and \code{issues}
(a possibly empty list of import issue records).
}
\description{
Helper to convert transmission date fields from character to Date if they're valid dates
}
\keyword{internal}
