% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattingHelpers.R
\name{.title_case_field}
\alias{.title_case_field}
\title{Simple title-casing for arbitrary field names (e.g. "date_first_transfer")}
\usage{
.title_case_field(x)
}
\description{
Simple title-casing for arbitrary field names (e.g. "date_first_transfer")
}
\keyword{internal}
