% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_dataset_exists}
\alias{.tv_block_dataset_exists}
\title{Does the DTA contain a dataset by this name (any kind)?}
\usage{
.tv_block_dataset_exists(dta, name)
}
\description{
Does the DTA contain a dataset by this name (any kind)?
}
\keyword{internal}
