% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_authorized_corrections}
\alias{.tv_block_render_authorized_corrections}
\title{\code{{AUTHORIZED_CORRECTIONS_LIST}} block: authorized-for-corrections names as a
bulleted list}
\usage{
.tv_block_render_authorized_corrections(doc, dta, arg)
}
\description{
\code{{AUTHORIZED_CORRECTIONS_LIST}} block: authorized-for-corrections names as a
bulleted list
}
\keyword{internal}
