% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_column_specs}
\alias{.tv_block_render_column_specs}
\title{\code{{COLUMN_SPECS}} block: a column-specification table per tabular dataset}
\usage{
.tv_block_render_column_specs(doc, dta, arg)
}
\description{
\code{{COLUMN_SPECS}} block: a column-specification table per tabular dataset
}
\keyword{internal}
