% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_contacts}
\alias{.tv_block_render_contacts}
\title{\code{{SUPPLIER_CONTACTS_TABLE}} / \code{{RECEIVER_CONTACTS_TABLE}} block: contacts for
one side of the agreement as a table}
\usage{
.tv_block_render_contacts(doc, dta, arg, side)
}
\description{
\code{{SUPPLIER_CONTACTS_TABLE}} / \code{{RECEIVER_CONTACTS_TABLE}} block: contacts for
one side of the agreement as a table
}
\keyword{internal}
