% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_signatures}
\alias{.tv_block_render_signatures}
\title{\code{{SIGNATURES_TABLE}} block: the approval and signatures table}
\usage{
.tv_block_render_signatures(doc, dta, arg)
}
\description{
\code{{SIGNATURES_TABLE}} block: the approval and signatures table
}
\keyword{internal}
