% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_version_history}
\alias{.tv_block_render_version_history}
\title{\code{{VERSION_HISTORY_TABLE}} block: version history as a table}
\usage{
.tv_block_render_version_history(doc, dta, arg)
}
\description{
\code{{VERSION_HISTORY_TABLE}} block: version history as a table
}
\keyword{internal}
