% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_takes_arg}
\alias{.tv_block_takes_arg}
\title{Block tokens that accept a \verb{:DATASET} argument}
\usage{
.tv_block_takes_arg()
}
\description{
Block tokens that accept a \verb{:DATASET} argument
}
\keyword{internal}
