% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_constraint_to_text}
\alias{.tv_constraint_to_text}
\title{Render one group_condition constraint to plain text}
\usage{
.tv_constraint_to_text(cst)
}
\description{
Render one group_condition constraint to plain text
}
\keyword{internal}
