% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_contacts_block}
\alias{.tv_contacts_block}
\title{Render a detailed contact block with signature lines for signers}
\usage{
.tv_contacts_block(org)
}
\description{
Render a detailed contact block with signature lines for signers
}
\keyword{internal}
