% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_ensure_run_props}
\alias{.tv_ensure_run_props}
\title{Ensure a run has a rPr child and return it}
\usage{
.tv_ensure_run_props(run_node)
}
\description{
Ensure a run has a rPr child and return it
}
\keyword{internal}
