% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_find_group_rule}
\alias{.tv_find_group_rule}
\title{Find a group_condition rule object by id within a DTA}
\usage{
.tv_find_group_rule(dta, rule_id)
}
\description{
Find a group_condition rule object by id within a DTA
}
\keyword{internal}
