% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_is_block_token}
\alias{.tv_is_block_token}
\title{Is this brace token a block placeholder?}
\usage{
.tv_is_block_token(token)
}
\description{
Is this brace token a block placeholder?
}
\keyword{internal}
