% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateRegions.R
\name{.tv_is_region_marker}
\alias{.tv_is_region_marker}
\title{Is this brace token a region marker (open or close)?}
\usage{
.tv_is_region_marker(token)
}
\description{
Is this brace token a region marker (open or close)?
}
\keyword{internal}
