% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_looks_markdown}
\alias{.tv_looks_markdown}
\title{Heuristic: does text look like markdown syntax?}
\usage{
.tv_looks_markdown(text)
}
\description{
Heuristic: does text look like markdown syntax?
}
\keyword{internal}
