% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_needs_yaml_style}
\alias{.tv_needs_yaml_style}
\title{Should a substituted placeholder be styled as embedded YAML code?}
\usage{
.tv_needs_yaml_style(keys)
}
\description{
Should a substituted placeholder be styled as embedded YAML code?
}
\keyword{internal}
