% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_parse_column_meta}
\alias{.tv_parse_column_meta}
\title{Parse the bracket metadata of a column bullet into named fields}
\usage{
.tv_parse_column_meta(meta)
}
\description{
Parse the bracket metadata of a column bullet into named fields
}
\keyword{internal}
