% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_process_information_block}
\alias{.tv_process_information_block}
\title{Render process information block (transmission, handling, corrections)}
\usage{
.tv_process_information_block(meta)
}
\description{
Render process information block (transmission, handling, corrections)
}
\keyword{internal}
