% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateRegions.R
\name{.tv_region_markers}
\alias{.tv_region_markers}
\title{Brace-delimited markers that open and close a repeating region}
\usage{
.tv_region_markers()
}
\description{
Brace-delimited markers that open and close a repeating region
}
\keyword{internal}
