% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_set_or_add_rpr_child}
\alias{.tv_set_or_add_rpr_child}
\title{Upsert a single run-property child element with attributes}
\usage{
.tv_set_or_add_rpr_child(rpr_node, child_name, attrs)
}
\description{
Upsert a single run-property child element with attributes
}
\keyword{internal}
