% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_set_preserve_space}
\alias{.tv_set_preserve_space}
\title{Set xml:space="preserve" on a text node, ignoring failures}
\usage{
.tv_set_preserve_space(node)
}
\description{
Set xml:space="preserve" on a text node, ignoring failures
}
\keyword{internal}
