% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_set_run_text}
\alias{.tv_set_run_text}
\title{Set run content from text, preserving tabs/newlines as Word elements}
\usage{
.tv_set_run_text(t_node, text)
}
\description{
Set run content from text, preserving tabs/newlines as Word elements
}
\keyword{internal}
