% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_split_word_chunks}
\alias{.tv_split_word_chunks}
\title{Split text into WordprocessingML chunks (text, line break, tab)}
\usage{
.tv_split_word_chunks(text)
}
\description{
Split text into WordprocessingML chunks (text, line break, tab)
}
\keyword{internal}
