% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateDocx.R
\name{.tv_version_history}
\alias{.tv_version_history}
\title{Summarize a version-history list as a single line}
\usage{
.tv_version_history(version_history)
}
\description{
Summarize a version-history list as a single line
}
\keyword{internal}
