% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingValidation.R
\name{dta_group_stream_init}
\alias{dta_group_stream_init}
\title{Start Accumulating a Grouped Rule}
\usage{
dta_group_stream_init(rule)
}
\arguments{
\item{rule}{A grouped rule.}
}
\value{
A mutable accumulator.
}
\description{
Start Accumulating a Grouped Rule
}
\keyword{internal}
