% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingValidation.R
\name{dta_rule_stream_init}
\alias{dta_rule_stream_init}
\title{Start Accumulating a Rule Across Batches}
\usage{
dta_rule_stream_init(rule)
}
\arguments{
\item{rule}{A rule object.}
}
\value{
A mutable accumulator.
}
\description{
Start Accumulating a Rule Across Batches
}
\keyword{internal}
