% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingValidation.R
\name{dta_rule_stream_kind}
\alias{dta_rule_stream_kind}
\title{How a Rule Can Be Streamed}
\usage{
dta_rule_stream_kind(rule)
}
\arguments{
\item{rule}{A rule object.}
}
\value{
One of \code{"decomposable"}, \code{"keyed"}, \code{"grouped"} or \code{"unsupported"}.
}
\description{
How a Rule Can Be Streamed
}
\keyword{internal}
