% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingValidation.R
\name{dta_table_is_lazy}
\alias{dta_table_is_lazy}
\title{Is This Table Lazy?}
\usage{
dta_table_is_lazy(x)
}
\arguments{
\item{x}{A table representation.}
}
\value{
\code{TRUE} when the table is scanned rather than held in memory.
}
\description{
Is This Table Lazy?
}
\keyword{internal}
