% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingValidation.R
\name{dta_unique_columns}
\alias{dta_unique_columns}
\title{Columns Forming a Uniqueness Key}
\usage{
dta_unique_columns(rule)
}
\arguments{
\item{rule}{A uniqueness rule.}
}
\value{
A character vector of column names.
}
\description{
Columns Forming a Uniqueness Key
}
\keyword{internal}
