% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R
\name{DTAColumnSpec}
\alias{DTAColumnSpec}
\title{DTA Column Format Class}
\usage{
DTAColumnSpec(
  id,
  label = NULL,
  type = NULL,
  format = NULL,
  length = NULL,
  nullable = NULL,
  pattern = NULL,
  values = NULL,
  examples = NULL,
  description = NULL,
  colclass = NULL
)
}
\arguments{
\item{id}{Character. The id of the column.}

\item{label}{Character or NA. The label of the column.}

\item{type}{Character or NA. The type of the column.}

\item{format}{Character or NA. The format of the column.}

\item{length}{Numeric or NA. The max character length.}

\item{nullable}{Logical or NA. Whether the column can be null.}

\item{pattern}{Character or NA. The pattern of the column.}

\item{values}{Any or NA. The values of the column.}

\item{description}{Character or NA. The description of the column.}
}
\value{
An object of class DTAColumnSpec.
}
\description{
Class for column format.

This class defines the format of a column in a DTA dataset.
}
\examples{
col_format <- DTAColumnSpec(id = "STUDYID", type = "Char", nullable = FALSE, values = "1234-1234")
}
