% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{DTADataSet}
\alias{DTADataSet}
\title{DTADataSet Class}
\usage{
DTADataSet(
  name,
  type,
  files = list(),
  description = NULL,
  template_source = NULL,
  template_version = NULL,
  template_date = NULL
)
}
\arguments{
\item{name}{Character. Name of the container.}

\item{type}{Character. Dataset type: either \code{"tabular"}, whose files
are read and validated against column specifications, or \code{"file"},
whose files are only checked for presence, readability and non-emptiness.}

\item{files}{a list of DTAFile objects specifying input file information.}

\item{description}{Character or NA. Free-text description of the dataset.}

\item{template_source}{Character or NA. Source of the template used to
generate the dataset specification.}

\item{template_version}{Character or NA. Version of the template used.}

\item{template_date}{Character or NA. Date of the template used.}
}
\value{
An object of class DTADataSet.
}
\description{
Class to handle data from files
}
\examples{
ds <- DTADataSet(
  name = "example_dataset",
  type = "file",
  files = list(create_example_DTAFileCSV())
)
ds
}
