% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{DTADataSetTabular}
\alias{DTADataSetTabular}
\title{DTADataSetTabular Class}
\usage{
DTADataSetTabular(
  name,
  specs,
  tables = list(),
  files = list(),
  description = NULL,
  template_source = NULL,
  template_version = NULL,
  template_date = NULL
)
}
\arguments{
\item{name}{Character. Name of the container.}

\item{specs}{A DTAColumnSpecCollection object specifying the column specs.}

\item{tables}{List. A named list of tables to be validated and included in the DTADataSetTabular object.}

\item{files}{a list of DTAFile objects specifying input file information.}
}
\value{
An object of class DTADataSetTabular
}
\description{
Handles tabular data with column specifications and rules.
}
\examples{
\dontrun{
# Create sample tables
table1 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("V03", "V03", "EOT"))
table2 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("EOT", "V05", "EOT"))

# List of tables
tables <- list(table1 = table1, table2 = table2)

# Create the DTADataSetTabular object
data_obj <- DTADataSetTabular(DTAColumnSpecCollection, tables)
}
}
