% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFunctions.R
\name{exportColumnValueTable}
\alias{exportColumnValueTable}
\title{Generate table containing all potential values of a column}
\usage{
exportColumnValueTable(
  DTAColumnSpecCollection,
  file = "column_value_table.docx",
  id
)
}
\arguments{
\item{DTAColumnSpecCollection}{DTAColumnSpecCollection object containing column specifications.}

\item{file}{Character. The name of the word file, to which the table shall be written. Default is "column_value_table.docx".}

\item{id}{Character. The id of the column for which a table with all its values shall be generated.}
}
\value{
None. The function creates a Word document.
}
\description{
This function takes all values defined in a column and prints a word table containing those values. Can be copied into the DTA
}
\examples{

exportColumnValueTable(DTAColumnSpecCollection, "column_value_table.docx")
}
