% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{files-DTADataSet}
\alias{files-DTADataSet}
\alias{files}
\title{Get files from DTADataSet Object}
\usage{
files(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSet.}

\item{...}{void}

\item{name}{Optional single character or single integer. if NULL, returns a
list of all files. If character, returns the datasets with the specified name.
If integer, returns the datasets at the specified index.}
}
\value{
A DTAColumnSpecCollection object.
}
\description{
Method to get files from DTADataSet object.
}
\examples{
library(DTAtools)
ds <- create_example_DTADataSetTabular()
files(ds)
}
