% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-helpers.R
\name{get_authorized_for_corrections}
\alias{get_authorized_for_corrections}
\title{Get Authorized Contacts for Corrections}
\usage{
get_authorized_for_corrections(x)
}
\arguments{
\item{x}{A DTAMetaData object}
}
\value{
A character vector of names or list of contacts authorized for corrections
}
\description{
Retrieve list of contacts authorized to request corrections from receiver
}
\examples{
md <- create_example_DTAMetaData(2)
get_authorized_for_corrections(md)

}
