% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{get_column}
\alias{get_column}
\title{Get Column by ID Method}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}

\item{id}{Character. The ID of the column to retrieve.}
}
\value{
A DTAColumnSpec object corresponding to the specified ID.
}
\description{
Method to get a column format by its ID from the collection.
}
\examples{
\dontrun{
 column_format <- get_column(collection, "STUDYID")
}
}
