% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{get_data-DTAContainer}
\alias{get_data-DTAContainer}
\title{Get table from DTAContainer Object}
\arguments{
\item{x}{An object of class DTAContainer.}

\item{id}{Character or numeric. Name or index of the table to retrieve.}
}
\value{
A data.frame.
}
\description{
Extract a table from the tables in a DTAContainer object.
}
\examples{
\dontrun{
get_data(container)           # returns first table
get_data(container, "lab")   # returns table named "lab"
}
}
