% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-helpers.R
\name{get_receiver_reviewers}
\alias{get_receiver_reviewers}
\title{Get Receiver Contacts with Reviewer Role}
\usage{
get_receiver_reviewers(x, name_only = TRUE)
}
\arguments{
\item{x}{A DTAMetaData object}

\item{name_only}{Logical. If TRUE, return only names; if FALSE, return full contact objects}
}
\value{
A character vector of reviewer names or list of reviewer contact objects
}
\description{
Retrieve receiver contacts who have the reviewer role (authorized to request corrections)
}
\examples{
md <- create_example_DTAMetaData(2)
get_receiver_reviewers(md)

}
