% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{get_rules-DTAColumnSpecCollection}
\alias{get_rules-DTAColumnSpecCollection}
\title{Get Rules}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}
}
\value{
A list with rules defined
}
\description{
Method to get Rules from DTAColumnSpecCollection
}
\examples{
\dontrun{
 get_rules(collection)
}
}
