% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{getColumn-DTAContainer}
\alias{getColumn-DTAContainer}
\title{Get Column by ID Method}
\arguments{
\item{x}{An object of class DTAContainer}

\item{id}{Character. The ID of the column to retrieve.}
}
\value{
A DTAColumnSpec object corresponding to the specified ID.
}
\description{
Method to get a column format by its ID from the collection.
}
\examples{
column_format <- getColumn(dtadata, "STUDYID")
}
