% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{getColumnIds-DTAColumnSpecCollection}
\alias{getColumnIds-DTAColumnSpecCollection}
\alias{getColumnIds}
\title{Get Names Method}
\usage{
getColumnIds(x, ...)
}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}

\item{...}{void}
}
\value{
A character vector of column names.
}
\description{
Method to get the names of columns in the collection.
}
\examples{
names <- getColumnIds(collection)
}
