% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{getContainer-DTA}
\alias{getContainer-DTA}
\title{Get container}
\arguments{
\item{x}{An object of class DTA.}

\item{name}{Optional character vector. One or more container names to retrieve.}
}
\value{
Either a list of DTRContainer objects s or a single named DTAContainer.
}
\description{
Method to get one or more containers from a DTA object.
}
\examples{
getContainer(DTAContainer)
getContainer(DTAContainer, "container1")
getContainer(DTAContainer, c("container1", "container2"))
}
