% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{labels-DTADataSet}
\alias{labels-DTADataSet}
\alias{labels}
\title{List of data labels within DTADataSet Object}
\usage{
labels(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSet}

\item{...}{void}
}
\value{
A vector
}
\description{
Method to get a all data labels within a DTADataSet Object.
}
\examples{
\dontrun{
labels <- labels(dtadata)
}
}
