% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleCollection-class.R
\name{length-DTARuleCollection}
\alias{length-DTARuleCollection}
\title{length method for DTARuleCollection}
\arguments{
\item{x}{A DTARuleCollection object}
}
\value{
Integer. The number of rules in the collection
}
\description{
Returns the number of rules in the collection
}
\examples{
library(DTAtools)
collection <- create_example_DTARuleCollection()
length(collection)
}
