% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{load_file-DTADataSet}
\alias{load_file-DTADataSet}
\title{loads file}
\arguments{
\item{x}{An object of class DTADataSet}
}
\value{
object of class DTADataSet with loaded data
}
\description{
Load the content of the file into dataset
}
\examples{
\dontrun{
column_format <- min_number_of_files(dtafiles)
}
}
