% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{makeDTAContainer}
\alias{makeDTAContainer}
\title{Generate DTAContainer Object}
\usage{
makeDTAContainer(specs, data)
}
\arguments{
\item{specs}{A DTAColumnSpecCollection object specifying the column specs.}

\item{data}{List. A list of tables to be validated and included in the DTAContainer object.}
}
\value{
An object of class DTAContainer.
}
\description{
This function generates a DTAContainer object from a DTAColumnSpecCollection object and a list of tables.
}
\examples{
#'
\dontrun{
# Create sample tables
table1 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("V03", "V03", "EOT"))
table2 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("EOT", "V05", "EOT"))

# List of tables
data <- list(table1 = table1, table2 = table2)

# Create the DTAContainer object
data_obj <- makeDTAContainer(specs = NULL, data)
}
}
