% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{matches_filename-DTAFileInfo}
\alias{matches_filename-DTAFileInfo}
\title{Matches Filename}
\arguments{
\item{file}{A character string representing the name of the file to check.}
}
\value{
A logical value indicating whether the filename matches the pattern.
}
\description{
Check if a filename matches a given pattern

Determines whether the provided filename matches the specified pattern.
}
\examples{
\dontrun{
matches_filename(dtafileinfo, "exact_file_name.tsv")
matches_filename(dtafileinfo, "\\\\.tsv$")
}
}
