% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{metadata-DTA}
\alias{metadata-DTA}
\title{Get Metadata}
\arguments{
\item{x}{An object of class DTA}
}
\value{
A list with metadata information
}
\description{
Method to get Metadata from DTA.
}
\examples{
\dontrun{
  get_metadata(DTA)
}
}
