% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{metadata}
\alias{metadata}
\title{Get Metadata}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}
}
\value{
A list with metadata information
}
\description{
Method to get Metadata from DTAColumnSpecCollection
}
\examples{
\dontrun{
 metadata(collection)
}
}
