% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{metadata-DTAContainer}
\alias{metadata-DTAContainer}
\title{Get Metadata}
\arguments{
\item{x}{An object of class DTAContainer}
}
\value{
A list with metadata information
}
\description{
Method to get Metadata from DTAContainer.
}
\examples{
\dontrun{
get_metadata(DTAContainer)
}
}
