% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{metadata}
\alias{metadata}
\title{Get Metadata}
\arguments{
\item{x}{An object of class DTADataSetTabular}
}
\value{
A list with metadata information
}
\description{
Method to get Metadata from DTADataSet.
}
\examples{
\dontrun{
metadata(DTADataSetTabular)
}
}
