% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{min_number_of_files-DTADataSet}
\alias{min_number_of_files-DTADataSet}
\title{get min number of files}
\arguments{
\item{x}{An object of class DTADataSet}
}
\value{
numeric: number of files
}
\description{
Returns the sum of min number of files specified all associated DTAFile
objects.
}
\examples{
ds <- DTADataSet(
  name = "example_dataset",
  type = "file",
  files = list(create_example_DTAFileCSV())
)
min_number_of_files(ds)
}
