% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{number_of_files}
\alias{number_of_files}
\title{Get number of files}
\arguments{
\item{x}{An object of class \code{DTAFileInfo}.}
}
\value{
The number of files.
}
\description{
Returns the number of files specified in a \code{DTAFileInfo} object.
}
\section{Methods}{

\describe{
\item{\code{DTAFileInfo}}{Returns the sum of the \code{number_of_files} property.}
}
}

\examples{
\dontrun{
  file_info <- DTAFileInfo("file.txt", number_of_files = 1)
  number_of_files(file_info)
}

}
