% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{numberOfFiles-DTAFileInfo}
\alias{numberOfFiles-DTAFileInfo}
\title{get max number of files}
\arguments{
\item{x}{An object of class DTAContainer}
}
\value{
number of files
}
\description{
Returns the number of files specified in the DTAFileInfo object.
}
\examples{
\dontrun{
column_format <- numberOfFiles(dtafileinfo)
}
}
