% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{print_dataset_info}
\alias{print_dataset_info}
\title{Print Dataset Information}
\usage{
print_dataset_info(x, ...)
}
\arguments{
\item{x}{A \code{DTADataSet} object whose information is to be printed.}
}
\value{
No return value. This function is called for its side effects (printing to the console).
}
\description{
Prints information about a \code{DTADataSet} object, including template source, version, date, and file information.
}
\details{
This method displays the template source, version, and date if available. It also summarizes the file information entries, indicating if none are present.
}
\examples{
# Assuming 'ds' is a DTADataSet object:
print_dataset_info(ds)

}
\seealso{
\code{\link{DTADataSet}}
}
