% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFile-class.R, R/DTAFileTabular-class.R
\name{print_file_info}
\alias{print_file_info}
\title{Print Information About a DTAFile Object}
\usage{
print_file_info(x, ...)
}
\arguments{
\item{x}{A \code{DTAFile} object whose information is to be printed.}
}
\value{
The input object \code{x}, returned invisibly.

The input object \code{x}, returned invisibly.
}
\description{
This method prints detailed information about a \code{DTAFile} object, including its filename, pattern, and the number of files associated with it. The information is displayed using the \code{cli} package for formatted output.

This method prints detailed information about a \code{DTAFile} object, including its filename, pattern, and the number of files associated with it. The information is displayed using the \code{cli} package for formatted output.
}
\details{
The function displays the filename and pattern of the \code{DTAFile} object. It also prints the minimum and maximum number of files, or a single value if both are equal.

The function displays the filename and pattern of the \code{DTAFile} object. It also prints the minimum and maximum number of files, or a single value if both are equal.
}
\examples{
\dontrun{
dta_file <- DTAFile(filename = "data.csv", pattern = "*.csv", min_number_of_files = 1, max_number_of_files = 1)
print_file_info(dta_file)
}

\dontrun{
dta_file <- DTAFile(filename = "data.csv", pattern = "*.csv", min_number_of_files = 1, max_number_of_files = 1)
print_file_info(dta_file)
}

}
\seealso{
\code{\link{DTAFile}}

\code{\link{DTAFile}}
}
