% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{read_dta_dataset_from_yaml}
\alias{read_dta_dataset_from_yaml}
\title{Read DTADataSet from YAML}
\usage{
read_dta_dataset_from_yaml(file)
}
\arguments{
\item{file}{Path to the YAML file containing DTADataSet specification}
}
\value{
An object of class DTADataSet
}
\description{
Constructs a DTADataSet object from a YAML file specification.
}
\examples{
require(DTAtools)
file <- system.file("extdata", "gf_dataset.yaml", package = "DTAtools")
dta <- read_dta_dataset_from_yaml(file)
}
