% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{read_dta_from_list}
\alias{read_dta_from_list}
\title{Read DTA from List}
\usage{
read_dta_from_list(x)
}
\arguments{
\item{x}{list}
}
\value{
An object of class DTA
}
\description{
Constructs a DTA object from a list.
}
\examples{
require(DTAtools)
file <- system.file("extdata", "clinical_dta.yaml", package = "DTAtools")
yaml_data <- yaml::read_yaml(file)
dta_obj <- read_dta_from_yaml(yaml_data)

}
