% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{read_dta_from_yaml}
\alias{read_dta_from_yaml}
\title{Read DTA from YAML}
\usage{
read_dta_from_yaml(file)
}
\arguments{
\item{file}{Path to the YAML file containing DTA specification}
}
\value{
An object of class DTA
}
\description{
Constructs a DTA object from a YAML file specification.
}
\examples{
file <- system.file("extdata", "clinical_dta.yaml", package = "DTAtools")
dta <- read_dta_from_yaml(file)
}
