% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{read_file-DTAFileInfo}
\alias{read_file-DTAFileInfo}
\alias{read_file}
\title{Read File for DTAFileInfo Objects}
\usage{
read_file(x, ...)
}
\arguments{
\item{x}{A DTAFileInfo object}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
NULL (not implemented)
}
\description{
Prototype function for reading files.
}
