% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileDelim-class.R
\name{read_file_execution-DTAFileDelim}
\alias{read_file_execution-DTAFileDelim}
\title{Read File for DTAFileDelim Objects}
\arguments{
\item{x}{A \code{DTAFileDelim} object containing file reading parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
A tibble containing the contents of the file if the filename
matches; otherwise, returns \code{NULL}.
}
\description{
Reads a TSV file using the parameters specified in a
\code{DTAFileDelim} object. This method uses \code{arrow::read_delim_arrow}
for efficient TSV parsing.
}
\seealso{
\code{\link{arrow::read_tsv_arrow}}
}
