% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfoDelim-class.R
\name{read_file_execution-DTAFileInfoDelim}
\alias{read_file_execution-DTAFileInfoDelim}
\title{Read File for DTAFileInfoDelim Objects}
\arguments{
\item{x}{A \code{DTAFileInfoDelim} object containing file reading parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
A tibble containing the contents of the file if the filename
matches; otherwise, returns \code{NULL}.
}
\description{
Reads a Delim file using the parameters specified in a
\code{DTAFileInfoDelim} object. This method uses \code{readr::read_Delim}
for efficient Delim parsing.
}
