% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfoTabular-class.R
\name{read_file_execution-DTAFileInfoTabular}
\alias{read_file_execution-DTAFileInfoTabular}
\title{Read File for DTAFileInfoTabular Objects}
\arguments{
\item{x}{A \code{DTAFileInfoTabular} object containing file reading parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
A tibble containing the contents of the file if the filename
matches; otherwise, returns \code{NULL}.
}
\description{
\code{DTAFileInfoTabular} is a virtual class. This method needs to be
implemented in derived classes like \code{DTAFileInfoTSV},
\code{DTAFileInfoCSV} or \code{DTAFileInfoDelim}.
}
