% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileTabular-class.R
\name{read_file_execution-DTAFileTabular}
\alias{read_file_execution-DTAFileTabular}
\title{Read File for DTAFileTabular Objects}
\usage{
## S7 method for class <DTAtools::DTAFile>
read_file_execution(x, file)
}
\arguments{
\item{x}{A \code{DTAFileTabular} object containing file reading parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
A tibble containing the contents of the file if the filename
matches; otherwise, returns \code{NULL}.
}
\description{
\code{DTAFileTabular} is a virtual class. This method needs to be
implemented in derived classes like \code{DTAFileTSV},
\code{DTAFileCSV} or \code{DTAFileDelim}.
}
