% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{rules-DTAContainer}
\alias{rules-DTAContainer}
\title{Get Rules}
\arguments{
\item{x}{An object of class DTAContainer}
}
\value{
A list with rules information
}
\description{
Method to get Rules from DTAContainer.
}
\examples{
\dontrun{
rules(DTAContainer)
}
}
