% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{single-bracket}
\alias{single-bracket}
\title{Extract multiple datasets with [}
\usage{
## S7 method for class <DTAtools::DTA>
x[i
}
\arguments{
\item{x}{An object of class \code{DTA}.}

\item{i}{A character vector of names or a numeric index vector.}
}
\value{
A named list of \code{DTADataSet} objects.
}
\description{
Extract one or more \code{DTADataSet} objects from a \code{DTA} object using
single-bracket notation. Always returns a named list.
}
\examples{
\dontrun{
  dta <- create_example_DTA()
  dta[c(1, 2)]
  dta[c("demographics", "vitals")]
}
}
