% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{specs_to_list}
\alias{specs_to_list}
\title{Transform DTAColumnSpecCollection to List}
\usage{
specs_to_list(DTAColumnSpecCollection)
}
\arguments{
\item{DTAColumnSpecCollection}{An object of class DTAColumnSpecCollection.}
}
\value{
List.
}
\description{
This function transforms a DTAColumnSpecCollection object to a list.
}
\examples{
\dontrun{
# Create a DTAColumnSpecCollection object
DTAColumnSpecCollection <- import_specs_from_yaml("path/to/yaml/file.yaml")

# Write the DTAColumnSpecCollection object to a YAML file
DTAColumnSpecCollectionToList(DTAColumnSpecCollection)
}
}
