% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleCollection-class.R
\name{[[-DTARuleCollection}
\alias{[[-DTARuleCollection}
\title{[[ operator for DTARuleCollection}
\arguments{
\item{x}{A DTARuleCollection object}

\item{i}{Index or name}
}
\value{
The selected rule(s) from the rules list
}
\description{
Returns the content of the rules list
}
