% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{tables-DTADataSet}
\alias{tables-DTADataSet}
\alias{tables}
\title{Get tables from DTADataSet Object}
\usage{
tables(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSet.}

\item{...}{void}
}
\value{
A DTAColumnSpecCollection object.
}
\description{
Method to get tables from DTADataSet object.
}
\examples{
library(DTAtools)
ds <- create_example_DTADataSetTabular()
tables(ds)
}
