% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattingHelpers.R
\name{THEME_COLORS}
\alias{THEME_COLORS}
\title{Professional Formatting Helpers for Document Export}
\usage{
THEME_COLORS
}
\description{
Utility functions for creating professional document styling
}
\keyword{internal}
