% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecStructure-class.R
\name{to_json_schema_type}
\alias{to_json_schema_type}
\title{to_json_schema_type}
\arguments{
\item{x}{An object of class DTAColumnSpecStructure.}
}
\description{
Converts a DTAColumnSpecStructure to a JSON Schema type.
}
