% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFunctions.R
\name{validate_table}
\alias{validate_table}
\title{Validate Table Against JSON Schema}
\usage{
validate_table(DTAColumnSpecCollection, table)
}
\arguments{
\item{DTAColumnSpecCollection}{A DTAColumnSpecCollection object.}

\item{table}{A data.frame to validate.}
}
\value{
Transformed and checked table (a data.frame) if valid, aborts otherwise. If invalid, returns a list containing summarised and full error data frames.
}
\description{
Validates a data.frame against a JSON Schema using jsonvalidate. The table is split into smaller chunks for validation to avoid argument limits.
}
