% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-helpers.R
\name{validate_transmission_dates}
\alias{validate_transmission_dates}
\title{Validate Transmission Dates}
\usage{
validate_transmission_dates(x)
}
\arguments{
\item{x}{A DTAMetaData object}
}
\value{
A list with elements is_valid (logical) and messages (character vector)
}
\description{
Check that transmission dates are valid (either Date objects or non-empty character phrases)
}
\examples{
md <- create_example_DTAMetaData(2)
validate_transmission_dates(md)

}
