% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{validation_status}
\alias{validation_status}
\title{Validation Status for DTADataSetTabular}
\usage{
validation_status(x, ...)
}
\arguments{
\item{x}{A \'DTADataSetTabular\' object.}

\item{...}{Additional arguments:
\describe{
\item{tables}{NULL (default), character table names, or numeric table indices.}
}}
}
\value{
A data.frame with validation status per table.
}
\description{
Returns a compact status table for validated tables.
}
