% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{writeDTAColumnSpecCollectionToJson}
\alias{writeDTAColumnSpecCollectionToJson}
\title{Write DTAColumnSpecCollection to JSON File}
\usage{
writeDTAColumnSpecCollectionToJson(
  DTAColumnSpecCollection,
  file,
  pretty = TRUE
)
}
\arguments{
\item{DTAColumnSpecCollection}{An object of class DTAColumnSpecCollection.}

\item{file}{Character. Path to the JSON file to write the specifications to.}

\item{pretty}{Logical. Whether to pretty-print the JSON. Default is TRUE.}
}
\value{
NULL. The function writes the DTAColumnSpecCollection to a JSON file.
}
\description{
This function writes a DTAColumnSpecCollection object to a JSON file.
}
\examples{
\dontrun{
# Create a DTAColumnSpecCollection object
DTAColumnSpecCollection <- import_specs_from_yaml("path/to/yaml/file.yaml")

# Write the DTAColumnSpecCollection object to a JSON file
writeDTAColumnSpecCollectionToJson(DTAColumnSpecCollection, "path/to/output/file.json")
}
}
