% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{writeDTAColumnSpecCollectionToYaml}
\alias{writeDTAColumnSpecCollectionToYaml}
\title{Write DTAColumnSpecCollection to YAML File}
\usage{
writeDTAColumnSpecCollectionToYaml(DTAColumnSpecCollection, "path/to/output/file.yaml")
}
\arguments{
\item{DTAColumnSpecCollection}{An object of class DTAColumnSpecCollection.}

\item{file}{Character. Path to the YAML file to write the specifications to.}
}
\value{
NULL. The function writes the DTAColumnSpecCollection to a YAML file.
}
\description{
This function writes a DTAColumnSpecCollection object to a YAML file.
}
\examples{
\dontrun{
# Create a DTAColumnSpecCollection object
DTAColumnSpecCollection <- importDTAColumnSpecCollectionFromYaml("path/to/yaml/file.yaml")

# Write the DTAColumnSpecCollection object to a YAML file
writeDTAColumnSpecCollectionToYaml(DTAColumnSpecCollection, "path/to/output/file.yaml")
}
}
