% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{as.list}
\alias{as.list}
\title{as.list for DTAColumnSpecCollection}
\usage{
## S7 method for class <DTAtools::DTAColumnSpecCollection>
as.list(x, ...)
}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}

\item{...}{Additional arguments (not used)}
}
\value{
A list representation of the DTAColumnSpecCollection object.
}
\description{
This function transforms a DTAColumnSpecCollection object to a list.
}
\examples{
x <- create_example_DTAColumnSpecCollection()
as.list(x)
}
