% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecStructure-class.R
\name{as.list}
\alias{as.list}
\title{as.list method for as.list.DTAColumnSpecStructure}
\usage{
## S7 method for class <DTAtools::DTAColumnSpecStructure>
as.list(x, ...)
}
\arguments{
\item{x}{A DTAColumnSpecStructure object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A named list with the DTAColumnSpecStructure properties.
}
\description{
Converts a DTAColumnSpecStructure object to a list.
}
