% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColUnique-class.R
\name{as.list}
\alias{as.list}
\title{as.list for DTARuleColCondition}
\usage{
## S7 method for class <DTAtools::DTARuleColUnique>
as.list(x)
}
\arguments{
\item{x}{An object of class DTARuleColCondition}

\item{...}{Additional arguments (not used).}
}
\value{
A named list containing the properties of the DTARuleColCondition object.
}
\description{
Convert a DTARuleColCondition object to a list.
}
