% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{as.list.DTAColumnSpecCollection}
\alias{as.list.DTAColumnSpecCollection}
\title{as.list for DTAColumnSpecCollection}
\usage{
\method{as.list}{DTAColumnSpecCollection}(x, ...)
}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}

\item{...}{Additional arguments (not used)}
}
\value{
A list representation of the DTAColumnSpecCollection object.
}
\description{
This function transforms a DTAColumnSpecCollection object to a list.
}
\examples{
\dontrun{
# Create a DTAColumnSpecCollection object
x <- import_columns_from_yaml("path/to/yaml/file.yaml")

# Write the DTAColumnSpecCollection object to a YAML file
as.list(x)
}
}
