% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-helpers.R,
%   R/DTARuleCollection-class.R
\name{as.list}
\alias{as.list}
\alias{[[,DTAtools::DTARuleCollection-method}
\alias{length,DTAtools::DTARuleCollection-method}
\title{Convert DTAMetaData to List}
\usage{
## S7 method for class <DTAtools::DTAMetaData>
as.list(x, ...)

## S7 method for class <DTAtools::DTARuleCollection>
as.list(x)

## S7 method for class <DTAtools::DTARuleCollection>
x[[i

## S7 method for class <DTAtools::DTARuleCollection>
length(x)
}
\arguments{
\item{x}{A DTARuleCollection object}

\item{...}{Additional arguments (not used)}

\item{i}{Index or name}
}
\value{
A list with all metadata fields

list

The selected rule(s) from the rules list

Integer. The number of rules in the collection
}
\description{
Convert a DTAMetaData object to a nested list structure suitable for YAML export
or programmatic access.

returns list from DTARuleCollection object

Returns the content of the rules list

Returns the number of rules in the collection
}
\details{
as.list
}
\examples{
library(DTAtools)
md <- create_example_DTAMetaData(2)
md_list <- as.list(md)

library(DTAtools)
collection <- create_example_DTARuleCollection()
length(collection)
}
