% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-helpers.R
\name{as.list}
\alias{as.list}
\title{Convert DTAMetaData to List}
\usage{
## S7 method for class <DTAtools::DTAMetaData>
as.list(x, ...)
}
\arguments{
\item{x}{An object of class DTAMetaData}

\item{...}{Additional arguments (not used)}
}
\value{
A list with all metadata fields
}
\description{
Convert a DTAMetaData object to a nested list structure suitable for YAML export
or programmatic access.
}
\examples{
library(DTAtools)
md <- create_example_DTAMetaData(2)
md_list <- as.list(md)

}
