% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R,
%   R/DTAColumnSpecCollection-class.R, R/DTAColumnSpecStructure-class.R
\name{as_json_schema}
\alias{as_json_schema}
\title{as_json_schema}
\arguments{
\item{x}{A \code{DTAColumnSpecStructure} object.}
}
\value{
A list representing the JSON Schema.
}
\description{
Converts a DTAColumnSpecStructure to a JSON Schema.

Converts a DTAColumnSpec s into a JSON Schema.

Converts a DTAColumnSpecStructure to a JSON Schema.
}
\examples{
library(DTAtools)
specs <- create_example_DTAColumnSpecCollection()
as_json_schema(specs)
}
