% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R,
%   R/DTAColumnSpecStructure-class.R, R/DTAColumnSpecStructureSAS-class.R
\name{as_json_schema_type}
\alias{as_json_schema_type}
\title{as_json_schema_type}
\arguments{
\item{x}{An object of class DTAColumnSpecStructure.}
}
\description{
Converts a DTAColumnSpec to a JSON Schema type.

Converts a DTAColumnSpecStructure to a JSON Schema type.

Converts a DTAColumnSpecStructure to a JSON Schema type.
}
