% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARule-class.R,
%   R/DTARuleColCondition-class.R, R/DTARuleColRange-class.R,
%   R/DTARuleColUnique-class.R
\name{check}
\alias{check}
\title{check}
\usage{
check(x, ...)

## S7 method for class <DTAtools::DTARuleColCondition>
check(x, tab)

## S7 method for class <DTAtools::DTARuleColRange>
check(x, tab)

## S7 method for class <DTAtools::DTARuleColUnique>
check(x, tab)
}
\description{
check rule against data

check rule against data

check rule against data

check rule against data
}
\examples{
\dontrun{
 # do not manually create DTARule objects, use derived classes instead
}
\dontrun{
 # TODO
}
\dontrun{
 # TODO
}
\dontrun{
 # TODO
}
}
