% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{colspec,DTAtools::DTADataSetTabular-method}
\alias{colspec,DTAtools::DTADataSetTabular-method}
\title{Get Column by ID Method}
\usage{
## S7 method for class <DTAtools::DTADataSetTabular>
colspec(x, id)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}

\item{id}{Character. The ID of the column to retrieve.}
}
\value{
A DTAColumnSpec object corresponding to the specified ID.
}
\description{
Method to get a column format by its ID from the collection.
}
\examples{
ds <- create_example_DTADataSetTabular(2)
colspec(ds, "STUDYID")
}
