% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{column_preview}
\alias{column_preview}
\title{Preview Column IDs in a DTAColumnSpecCollection}
\usage{
column_preview(x, ...)
}
\arguments{
\item{x}{A \code{DTAColumnSpecCollection} object.}
}
\value{
A character string representing a preview of the column IDs.
}
\description{
Provides a preview of the column IDs contained within a \code{DTAColumnSpecCollection} object.
If the collection contains more than 5 specs, it returns the IDs of the first four specs,
an ellipsis, and the last column's ID. If there are 5 or fewer specs, it returns all column IDs.
}
\examples{
library(DTAtools)
x <- create_example_DTAColumnSpecCollection()
column_preview(x)
}
\seealso{
\code{\link{DTAColumnSpecCollection}}
}
