% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{container}
\alias{container}
\title{Get container}
\arguments{
\item{x}{An object of class DTA.}

\item{name}{Optional character vector. One or more container names to retrieve.}
}
\value{
Either a list of DTRContainer objects s or a single named DTAContainer.
}
\description{
Method to get one or more containers from a DTA object.
}
\examples{
\dontrun{
container(DTAContainer)
container(DTAContainer, "container1")
container(DTAContainer, c("container1", "container2"))
}
}
