% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{convertYamlToJson}
\alias{convertYamlToJson}
\title{Convert YAML Spec File to JSON}
\usage{
convertYamlToJson(yaml_file, json_file)
}
\arguments{
\item{yaml_file}{Character. Path to the YAML file.}

\item{json_file}{Character. Path to the output JSON file.}

\item{pretty}{Logical. Whether to pretty-print the JSON. Default is TRUE.}
}
\value{
NULL. Writes JSON to file.
}
\description{
Converts a YAML specification file to a JSON file.
}
\examples{
\dontrun{
  convertYamlToJson(yaml_file, json_file)
}
}
