% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{create_example_DTA}
\alias{create_example_DTA}
\title{Create Example DTA Object}
\usage{
create_example_DTA(title = "Example DTA", version = "1.0")
}
\arguments{
\item{title}{Character string. Title for the DTA object.}

\item{version}{Character string. Version of the DTA object.}
}
\value{
An object of class DTA with example data
}
\description{
Creates an example DTA object for demonstration purposes.
}
\examples{
\dontrun{
  example_dta <- create_example_DTA()
  print(example_dta)
}
}
