% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R
\name{create_example_DTAColumnSpec}
\alias{create_example_DTAColumnSpec}
\title{Create Example DTAColumnSpec}
\usage{
create_example_DTAColumnSpec(x, ...)
}
\description{
S7 method to create and return an example DTAColumnSpec object.
}
