% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{create_example_DTAContainer}
\alias{create_example_DTAContainer}
\title{Create Example DTAContainer}
\usage{
create_example_DTAContainer(index = 1)
}
\arguments{
\item{index}{Integer. Index of the example to create.}
}
\value{
An example DTAContainer object.
}
\description{
S7 method to create and return an example DTAContainer object.
}
\examples{
library(DTAtools)
create_example_DTAContainer()
}
