% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{create_example_DTADataSetTabular}
\alias{create_example_DTADataSetTabular}
\title{Create Example DTADataSetTabular}
\usage{
create_example_DTADataSetTabular(index = 1)
}
\arguments{
\item{index}{Integer. Index of the example to create.}
}
\value{
An example DTADataSetTabular object.
}
\description{
S7 method to create and return an example DTADataSetTabular object.
}
\examples{
library(DTAtools)
create_example_DTADataSetTabular()
}
