% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileCSV-class.R
\name{create_example_DTAFileCSV}
\alias{create_example_DTAFileCSV}
\title{Create Example DTAFileCSV Object}
\usage{
create_example_DTAFileCSV(index = 1)
}
\arguments{
\item{index}{example selector.}
}
\value{
An example \code{DTAFileCSV} object.
}
\description{
Creates an example \code{DTAFileCSV} object using example files.
}
\examples{
library(DTAtools)
create_example_DTAFileCSV()
}
