% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-class.R
\name{create_example_DTAMetaData}
\alias{create_example_DTAMetaData}
\title{Create Example DTAMetaData Object}
\usage{
create_example_DTAMetaData(index = 1)
}
\arguments{
\item{index}{Numeric. 1 = basic example, 2 = full example with history and transmission,
3 = example with date phrases for transfers}
}
\value{
An object of class \code{DTAMetaData} with example metadata.
}
\description{
This function creates example \code{DTAMetaData}
objects with various levels of completeness and structure.
}
\examples{
library(DTAtools)
example_metadata <- create_example_DTAMetaData()
print(example_metadata)
}
