% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-class.R
\name{create_example_DTAMetaData}
\alias{create_example_DTAMetaData}
\title{Create Example DTAMetaData Object}
\usage{
create_example_DTAMetaData(index = 1)
}
\value{
An object of class \code{DTAMetaData} with example metadata.
}
\description{
This function creates an example \code{DTAMetaData}
object with default values.
}
\examples{
library(DTAtools)
example_metadata <- create_example_DTAMetaData()
print(example_metadata)
}
