% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColCondition-class.R
\name{create_example_DTARuleColCondition}
\alias{create_example_DTARuleColCondition}
\title{create_example_DTARuleColCondition}
\usage{
create_example_DTARuleColCondition(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARule
}
\examples{
 library(DTAtools)
 create_example_DTARuleColCondition()
}
