% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColRange-class.R
\name{create_example_DTARuleColRange}
\alias{create_example_DTARuleColRange}
\title{create_example_DTARuleColRange}
\usage{
create_example_DTARuleColRange(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARuleColRange
}
\examples{
 library(DTAtools)
 create_example_DTARuleColRange()
}
