% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColUnique-class.R
\name{create_example_DTARuleColUnique}
\alias{create_example_DTARuleColUnique}
\title{create_example_DTARuleColUnique}
\usage{
create_example_DTARuleColUnique(index = 1)
}
\arguments{
\item{index}{rule selector}
}
\description{
create example for DTARuleColUnique
}
\examples{
library(DTAtools)
create_example_DTARuleColUnique()
}
