% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{datasets}
\alias{datasets}
\title{Get datasets}
\arguments{
\item{x}{An object of class DTA.}

\item{name}{Optional single character or single integer. if NULL, returns a
list of all datasetss. If character, returns the datasets with the specified name.
If integer, returns the datasets at the specified index.}
}
\value{
Either a list of DTADataSet objects or a single DTADataSet.
}
\description{
Method to get one or more datasetss from a DTA object.
}
\examples{
libary(DTAtools)
x <- create_example_DTA()
datasets(x)
datasets(x, "vitals")
datasets(x, 1)
}
