% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentBuilders.R
\name{.add_organization_section}
\alias{.add_organization_section}
\title{Add a full organization section (affiliation + individual contacts table)}
\usage{
.add_organization_section(doc, title, org_list)
}
\description{
Add a full organization section (affiliation + individual contacts table)
}
\keyword{internal}
