% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDocuments.R
\name{.convert_docx_to_pdf}
\alias{.convert_docx_to_pdf}
\title{Convert DOCX to PDF using pandoc or similar}
\usage{
.convert_docx_to_pdf(docx_file, pdf_file)
}
\description{
Convert DOCX to PDF using pandoc or similar
}
\keyword{internal}
