% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDocuments.R
\name{.organization_to_md_lines}
\alias{.organization_to_md_lines}
\title{Render an organization block (affiliation + individual contacts table) as Markdown lines}
\usage{
.organization_to_md_lines(title, org_list)
}
\description{
Render an organization block (affiliation + individual contacts table) as Markdown lines
}
\keyword{internal}
