% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-class.R
\name{.process_transmission_dates}
\alias{.process_transmission_dates}
\title{Process Transmission Dates}
\usage{
.process_transmission_dates(transmission)
}
\description{
Helper to convert transmission date fields from character to Date if they're valid dates
}
\keyword{internal}
