% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render}
\alias{.tv_block_render}
\title{Render one block placeholder at the cursor}
\usage{
.tv_block_render(doc, spec, dta)
}
\description{
Render one block placeholder at the cursor
}
\keyword{internal}
