% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_datasets}
\alias{.tv_block_render_datasets}
\title{\code{{DATASETS}} block: the full per-dataset section -- description, files,
columns, rules}
\usage{
.tv_block_render_datasets(doc, dta, arg)
}
\description{
\code{{DATASETS}} block: the full per-dataset section -- description, files,
columns, rules
}
\keyword{internal}
