% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_file_specs}
\alias{.tv_block_render_file_specs}
\title{\code{{FILE_SPECS}} block: an expected-files table per dataset}
\usage{
.tv_block_render_file_specs(doc, dta, arg)
}
\description{
\code{{FILE_SPECS}} block: an expected-files table per dataset
}
\keyword{internal}
