% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportTemplateBlocks.R
\name{.tv_block_render_validation_rules}
\alias{.tv_block_render_validation_rules}
\title{\code{{VALIDATION_RULES}} block: a validation-rule table per tabular dataset}
\usage{
.tv_block_render_validation_rules(doc, dta, arg)
}
\description{
\code{{VALIDATION_RULES}} block: a validation-rule table per tabular dataset
}
\keyword{internal}
