% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{double-bracket}
\alias{double-bracket}
\title{Extract single dataset with [[}
\usage{
## S7 method for class <DTAtools::DTA>
x[[i
}
\arguments{
\item{x}{An object of class \code{DTA}.}

\item{i}{A single character name or single numeric index.}
}
\value{
A single \code{DTADataSet} object.
}
\description{
Extract a single \code{DTADataSet} from a \code{DTA} object using
double-bracket notation.
}
\examples{
\dontrun{
  dta <- create_example_DTA()
  dta[[1]]
  dta[["demographics"]]
}
}
