% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{DTA}
\alias{DTA}
\title{DTA Class}
\usage{
DTA(container = NULL, metadata = NULL, ...)
}
\arguments{
\item{container}{A names list of DTAContainer objects}

\item{...}{arguments will be passed to DTAMetadata(...)}
}
\value{
An object of class DTA.
}
\description{
This class helps checking validity of data tables against transmission
specifications in data transfer agreements (DTA). Also it provides a way to
generate DTA/DTS documents from specifications.
}
\examples{

\dontrun{
# Create sample tables
table1 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("V03", "V03", "EOT"))
table2 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("EOT", "V05", "EOT"))

# List of tables
tables <- list(table1 = table1, table2 = table2)

# Create the DTAContainer object
data_obj <- DTAContainer(DTAColumnSpecCollection, tables)

DTA(container = list(data = data_obj))
}
}
