% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{DTA}
\alias{DTA}
\title{DTA Class}
\usage{
DTA(datasets = NULL, metadata = NULL, ...)
}
\arguments{
\item{datasets}{A named list of DTADataSet objects.}

\item{metadata}{A DTAMetaData object.}

\item{...}{If metadata is not set, additional arguments are passed to
DTAMetaData(...).}
}
\value{
An object of class DTA.
}
\description{
This class helps checking validity of data tables against transmission
specifications in data transfer agreements (DTA). Also it provides a way to
generate DTA/DTS documents from specifications.
}
\examples{

\dontrun{
# Create sample tables
table1 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("V03", "V03", "EOT"))
table2 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("EOT", "V05", "EOT"))

# List of tables
tables <- list(table1 = table1, table2 = table2)

# Create the DTADataSet object
data_obj <- DTADataSet(DTAColumnSpecCollection, tables)

DTA(datasets = list(data = data_obj))
}
}
