% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateRules.R
\name{dta_condition_violation_message}
\alias{dta_condition_violation_message}
\title{Message for a Conditional Rule Violation}
\usage{
dta_condition_violation_message(id, n)
}
\arguments{
\item{id}{Character. The rule id.}

\item{n}{Integer. Number of violating rows.}
}
\value{
A single string.
}
\description{
Message for a Conditional Rule Violation
}
\keyword{internal}
