% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{dta_dataset_from_list}
\alias{dta_dataset_from_list}
\title{DTADataSet from list}
\usage{
dta_dataset_from_list(x, recursive = TRUE)
}
\arguments{
\item{x}{List}

\item{recursive}{Logical, if TRUE (default) processes nested datasets}
}
\value{
An object of class DTADataSet
}
\description{
Constructs a DTADataSet object from a list
}
\examples{
require(DTAtools)
file <- system.file("extdata", "gf_dataset.yaml", package = "DTAtools")
yaml_data <- yaml::read_yaml(file)
dta_obj <- dta_dataset_from_list(yaml_data)
}
