% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{DTAColumnSpecCollection}
\alias{DTAColumnSpecCollection}
\title{DTAColumnSpecCollection Class}
\usage{
DTAColumnSpecCollection(specs, rules = list())
}
\arguments{
\item{specs}{A named list of DTAColumnSpec objects. The names should correspond to the ids of the specs}

\item{rules}{A list of DTARule object.}

\item{template}{A list of template information.}
}
\value{
An object of class DTAColumnSpecCollection.
}
\description{
This class represents a collection of DTAColumnSpec objects with optional template.
}
\examples{
col1 <- DTAColumnSpec(id = "STUDYID", type = "Char", nullable = TRUE)
col2 <- DTAColumnSpec(id = "VISIT", type = "Char", nullable = FALSE)
collection <- DTAColumnSpecCollection(specs = list(STUDIYID = col1, VISIT = col2))
}
