% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{DTAColumnSpecCollection}
\alias{DTAColumnSpecCollection}
\title{DTAColumnSpecCollection Class}
\usage{
DTAColumnSpecCollection(columns, rules = NULL)
}
\arguments{
\item{columns}{A named list of DTAColumnSpec objects. The names should correspond to the ids of the columns}

\item{rules}{A list of DTARule objects, or NULL.}
}
\value{
An object of class DTAColumnSpecCollection.
}
\description{
This class represents a collection of DTAColumnSpec objects with optional template.
}
\examples{
col1 <- DTAColumnSpec(id = "STUDYID", type = "SAS Char", nullable = TRUE)
col2 <- DTAColumnSpec(id = "VISIT", type = "SAS Char", nullable = FALSE)
collection <- DTAColumnSpecCollection(columns = list(STUDYID = col1, VISIT = col2))
}
