% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecStructure-class.R
\name{DTAColumnSpecStructure}
\alias{DTAColumnSpecStructure}
\title{DTA Column Spec Structure}
\usage{
DTAColumnSpecStructure(
  type = NULL,
  format = NULL,
  length = NULL,
  backend = NULL
)
}
\arguments{
\item{type}{Character or NA. The type of the column.}

\item{format}{Character or NA. The format of the column.}

\item{length}{Numeric or NA. The max character length.}
}
\description{
Class for column types.

This class defines the structure of a column in a DTA dataset.
}
\examples{
 DTAColumnSpecStructure(type = "Char", format = "$12.", length = 12, backend = "SAS")
}
