% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecStructureSAS-class.R
\name{DTAColumnSpecStructureSAS}
\alias{DTAColumnSpecStructureSAS}
\title{DTA Column Spec Structure SAS}
\usage{
DTAColumnSpecStructureSAS(type = NULL, format = NULL, length = NULL)
}
\arguments{
\item{type}{Character or NA. The type of the column.}

\item{format}{Character or NA. The format of the column.}

\item{length}{Numeric or NA. The max character length.}
}
\description{
Class for column types for SAS structures.

This class defines the structure of a column in a DTA dataset.
}
\examples{
\dontrun{
 DTAColumnSpecStructureSAS(type = "Char", format = "$12.", length = 12)
}
}
