% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{DTADataSet}
\alias{DTADataSet}
\title{DTADataSet Class}
\usage{
DTADataSet(name, specs, data = list(), fileinfo = list())
}
\arguments{
\item{name}{Character. Name of the container.}

\item{specs}{A DTAColumnSpecCollection object specifying the column specs.}

\item{data}{List. A list of arrow Table to be validated and included in the DTADataSet object.}

\item{fileinfo}{a list of DTAFileInfo objects specifying input file information.}
}
\value{
An object of class DTADataSet. If validation fails, returns a list containing summarised and full error data frames.
}
\description{
Handles tables  based on column specifications. Every table will be validated using the column specifications.
}
\examples{

\dontrun{
# Create sample tables
table1 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("V03", "V03", "EOT"))
table2 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("EOT", "V05", "EOT"))

# List of tables
tables <- list(table1 = table1, table2 = table2)

# Create the DTADataSet object
data_obj <- DTADataSet(DTAColumnSpecCollection, tables)
}
}
