% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{DTAContainer}
\alias{DTAContainer}
\title{DTAContainer Class}
\usage{
DTAContainer(specs, data)
}
\arguments{
\item{specs}{A DTAColumnSpecCollection object specifying the column specs.}

\item{data}{List. A list of tables to be validated and included in the DTAContainer object.}
}
\value{
An object of class DTAContainer. If validation fails, returns a list containing summarised and full error data frames.
}
\description{
Handles tables  based on column specifications. Every table will be validated using the column specifications.
}
\examples{

\dontrun{
# Create sample tables
table1 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("V03", "V03", "EOT"))
table2 <- data.frame(STUDYID = c("1234", "1234", "1234"), VISIT = c("EOT", "V05", "EOT"))

# List of tables
tables <- list(table1 = table1, table2 = table2)

# Create the DTAContainer object
data_obj <- DTAContainer(DTAColumnSpecCollection, tables)
}
}
