% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{DTADataSet}
\alias{DTADataSet}
\title{DTADataSet Class}
\usage{
DTADataSet(
  name,
  type,
  files = list(),
  description = NULL,
  template_source = NULL,
  template_version = NULL,
  template_date = NULL
)
}
\arguments{
\item{name}{Character. Name of the container.}

\item{type}{Character. Dataset type, must be one of \verb{__DTAtools_supported_dataset_types__}.}

\item{files}{a list of DTAFile objects specifying input file information.}
}
\value{
An object of class DTADataSet. If validation fails, returns a list containing summarised and full error data frames.
}
\description{
Class to handle data from files
}
\examples{

\dontrun{
# Create sample 
}
}
