% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helpers.R
\name{DTADataSetFactory}
\alias{DTADataSetFactory}
\title{Create a DTADataSetFactory Object}
\usage{
DTADataSetFactory(type, ...)
}
\arguments{
\item{type}{Character. The type specification, potentially prefixed with a backend identifier.}

\item{...}{Character. Arguments passed on to the specific backend constructor.}
}
\value{
An object derived from class \code{DTADataSet} like \code{DTADataSetTabular},
depending on the backend specified.
}
\description{
Constructs a DTADataSetFactory object for a specified backend (e.g., SAS or R),
based on the provided type.
}
\examples{
library(DTAtools)
DTAColumnSpecStructureFactory(type = "tabular", name = "mydataset")

}
\seealso{
\code{\link{DTAtools::DTADataSet}}, \code{\link{DTAtools::DTADataSetTabular}}
}
