% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helpers.R
\name{DTADataSetFactory}
\alias{DTADataSetFactory}
\title{Create a DTADataSetFactory Object}
\usage{
DTADataSetFactory(type, columns = NULL, rules = NULL, files = NULL, ...)
}
\arguments{
\item{type}{Character. The type specification, potentially prefixed with a backend identifier.}

\item{columns}{A list with column specifications (optional)}

\item{rules}{A list of rules specifications (optional)}

\item{files}{A list of for conversion to DTAFile objects (optional)}

\item{...}{Character. Arguments passed on to the specific backend constructor.}
}
\value{
An object derived from class \code{DTADataSet} like \code{DTADataSetTabular},
depending on the backend specified.
}
\description{
Constructs a DTADataSetFactory object for a specified backend (e.g., SAS or R),
based on the provided type.
}
\examples{
library(DTAtools)
DTADataSetFactory(
  type = "file",
  name = "mydataset",
  files = list(type = "csv", filename = "clinical_data.csv")
)

}
\seealso{
\code{\link{DTADataSet}}, \code{\link{DTADataSetTabular}}
}
