% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetFile-class.R
\name{DTADataSetFile}
\alias{DTADataSetFile}
\title{DTADataSetFile Class}
\usage{
DTADataSetFile(
  name,
  paths = character(),
  files = list(),
  description = NULL,
  template_source = NULL,
  template_version = NULL,
  template_date = NULL
)
}
\arguments{
\item{name}{Character. Name of the container.}

\item{paths}{Character vector of file paths to validate.}

\item{files}{A list of DTAFile objects specifying input file information.}

\item{description}{Character or NULL. Optional description.}

\item{template_source}{Character or NA. Source of the template used to
generate the dataset specification.}

\item{template_version}{Character or NA. Version of the template used.}

\item{template_date}{Character or NA. Date of the template used.}
}
\value{
An object of class DTADataSetFile.
}
\description{
Handles file-backed datasets that only need to verify that one
or more referenced files exist, are readable, and are not empty.
}
