% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFile-class.R
\name{DTAFile-class}
\alias{DTAFile-class}
\alias{DTAFile}
\title{DTAFile Class}
\usage{
DTAFile(
  filename,
  pattern = FALSE,
  pattern_description = NULL,
  number_of_files = NULL,
  min_number_of_files = NULL,
  max_number_of_files = NULL,
  info = NULL
)
}
\arguments{
\item{filename}{Character vector of file names or regular expression patterns
to match files.}

\item{pattern}{Logical; if \code{TRUE}, \code{filename} is treated as a regex
pattern. Default is \code{FALSE}.}

\item{pattern_description}{Character or \code{NULL}; human-readable
description of the \code{filename} pattern.}

\item{number_of_files}{Numeric or \code{NULL}; number of files
expected. Default is \code{1}. If two numbers are provided,
they represent min and maximum of files expected.}

\item{min_number_of_files}{Numeric or \code{NULL}; minimum number of files
expected.}

\item{max_number_of_files}{Numeric or \code{NULL}; maximum number of files
expected.}

\item{info}{Character or \code{NULL}; free-text description of the file.}
}
\value{
An object of class \code{DTAFile} containing file parsing
information.
}
\description{
The \code{DTAFile} class stores metadata and parsing instructions for
delimited data files. It specifies file names (or patterns), the expected
number of files, and how to read them, including separator, header presence,
quoting, and column types.
}
\details{
This class is used internally by the DTAtoolsR package to
manage metadata and properties of DTA files.
}
\examples{
file_info <- DTAFile("file.txt")
file_info_pattern <- DTAFile("file\\\\d+\\\\.txt", pattern = TRUE)
}
\keyword{internal}
