% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfoCSV-class.R
\name{DTAFileInfoCSV-class}
\alias{DTAFileInfoCSV-class}
\alias{DTAFileInfoCSV}
\title{DTAFileInfoCSV Class Constructor}
\usage{
DTAFileInfoCSV(
  filename,
  pattern = FALSE,
  number_of_files = 1,
  has_header = TRUE,
  quote = "\\""
)
}
\arguments{
\item{filename}{Character vector of file names or regular expression patterns
to match files.}

\item{pattern}{Logical; if \code{TRUE}, \code{filename} is treated as a regex
pattern. Default is \code{FALSE}.}

\item{number_of_files}{Numeric or \code{NULL}; maximum number of files
expected. Default is \code{1}.}

\item{has_header}{Logical; \code{TRUE} if the first row is a header. Default
is \code{TRUE}.}

\item{quote}{Character or \code{NULL}; quoting character for fields. Default
is \code{'"'}.}

\item{sep}{Character. Field separator used in the TSV file.
Defaults to tab ("\\t").}
}
\value{
An object of class \code{DTAFileInfoCSV}.
}
\description{
Defines the S7 class \code{DTAFileInfoCSV}, which extends \code{DTAFileInfo}
to represent metadata and configuration for CSV (Tab-Separated Values)
data files.
}
\examples{
 \dontrun{
}
}
\seealso{
\code{\link{DTAFileInfo}}
}
