% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileTabular-class.R
\name{DTAFileTabular-class}
\alias{DTAFileTabular-class}
\alias{DTAFileTabular}
\title{DTAFileTabular Class Constructor}
\usage{
DTAFileTabular(
  filename,
  pattern = FALSE,
  number_of_files = 1,
  missing_values = "",
  sep = "\\t",
  has_header = TRUE,
  quote = "\\"",
  encoding = "UTF-8"
)
}
\arguments{
\item{filename}{Character vector of file names or regular expression patterns
to match files.}

\item{pattern}{Logical; if \code{TRUE}, \code{filename} is treated as a regex
pattern. Default is \code{FALSE}.}

\item{number_of_files}{Numeric or \code{NULL}; maximum number of files
expected. Default is \code{1}.}

\item{sep}{Character. Field separator used in the TSV file.
Defaults to tab ("\\t").}

\item{has_header}{Logical; \code{TRUE} if the first row is a header. Default
is \code{TRUE}.}

\item{quote}{Character or \code{NULL}; quoting character for fields. Default
is \code{'"'}.}

\item{col_types}{Character string specifying the type of each column (e.g.,
\code{"cccidcl"}). Default is \code{NULL}.}
}
\value{
An object of class \code{DTAFileTabular}.
}
\description{
Defines the S7 class \code{C}, which extends \code{DTAFile}
to represent metadata and configuration for Delim (Tab-Separated Values)
data files.
}
\seealso{
\code{\link{DTAFile}}
}
