% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-class.R
\name{DTAMetaData}
\alias{DTAMetaData}
\title{DTAMetaData Class}
\usage{
DTAMetaData(name, version = NULL, author = NULL)
}
\arguments{
\item{version}{version}

\item{author}{author}
}
\value{
An object of class DTAMetaData.
}
\description{
This class holds metadata information
}
\examples{

\dontrun{
DTAMetaData(version = "1.0", author = "John Doe")
}
}
