% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-class.R
\name{DTAMetaData}
\alias{DTAMetaData}
\title{DTAMetaData Class}
\usage{
DTAMetaData(
  title,
  version = NULL,
  date = NULL,
  header = NULL,
  receiver = list(),
  supplier = list(),
  transmission = list()
)
}
\arguments{
\item{title}{title}

\item{version}{version}

\item{date}{date}

\item{header}{header}

\item{receiver}{receiver}

\item{supplier}{supplier}

\item{transmission}{transmission}
}
\value{
An object of class DTAMetaData.
}
\description{
This class holds metadata information
}
\examples{

DTAMetaData(title = "Clinical Data Transfer", version = "1.0")
}
