% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARule-class.R
\name{DTARule}
\alias{DTARule}
\title{DTARule Class}
\usage{
DTARule(id, type, description)
}
\arguments{
\item{id}{Character. A unique identifier for the rule.}

\item{type}{Character. The type of the rule (e.g., "check_range", "check_unique").}
}
\value{
An object of class \code{DTARule}.
}
\description{
Represents a single rule for validating data tables. The rule can be of various types,
such as \code{check_range}, \code{check_unique}, or \code{check_col_condition}
}
\examples{
rule <- DTARuleFactory("rule1", "col_unique", columns = "SUBJID")
rule
}
