% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARule-class.R
\name{DTARule}
\alias{DTARule}
\title{DTARule Class}
\usage{
DTARule(id, type, column = NULL, range = NULL, condition = NULL, then = NULL)
}
\arguments{
\item{id}{Character. A unique identifier for the rule.}

\item{type}{Character. The type of the rule (e.g., "check_range", "check_unique").}

\item{column}{List. Optional depending on the rule id. Used in check_range and check_unique. In check_unique it is a list of columns that will be collectively checked if the combinations are unique throughout the table}

\item{range}{Vector or List. Optional depending on the rule id. Will be used in check_range to check the range of the values in a given column.}

\item{condition}{List. Optional depending on the rule id. A list of conditions to check}

\item{then}{List. Optional depending on the rule id. A list of conditions to check - used with condition}
}
\value{
An object of class \code{DTARule}.
}
\description{
Represents a single rule for validating data tables. The rule can be of various types,
such as \code{check_range}, \code{check_unique}, or \code{check_condition}
}
\examples{
# Create a check_range rule
rule1 <- DTARule(
  id = "rule1",
  type = "check_range",
  column = "age",
  range = c(18, 65)
)

# Create a check_unique rule
rule2 <- DTARule(
  id = "rule2",
  type = "check_unique",
  column = "id"
)
}
