% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColCondition-class.R
\name{DTARuleColCondition}
\alias{DTARuleColCondition}
\title{DTARuleColCondition Class}
\usage{
DTARuleColCondition(id, description = NULL, condition = NULL, then = NULL)
}
\arguments{
\item{id}{Character. A unique identifier for the rule.}

\item{condition}{List. A list of conditions to check}

\item{then}{List. A list of conditions that must be true if the
conditions are met}
}
\value{
An object of class \code{DTARuleColCondition}.
}
\description{
A rule for validating data tables with conditions
}
\examples{
# Create a check_range rule
DTAtools::DTARuleColCondition(
 id = "check_gfreasnd2",
 condition = list(
   GFREASND = list(
     empty = FALSE
   )
 ),
 then = list(
   GFSTAT = list(
     empty = FALSE
   ),
   GFORRES = list(
     empty = TRUE
   )
 )
)
}
