% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColRange-class.R
\name{DTARuleColRange}
\alias{DTARuleColRange}
\title{DTARuleColRange Class}
\usage{
DTARuleColRange(
  id,
  columns = NULL,
  range = NULL,
  description = NULL,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{id}{Character. A unique identifier for the rule.}

\item{columns}{Charaterc. columns checked for unique combinations.}

\item{range}{Vector or List. Used in check_range to check value ranges in a
columns.}

\item{type}{Character. The rule type (e.g., "check_range", "check_unique").}
}
\value{
An object of class \code{DTARuleColRange}.
}
\description{
Represents a rule for checking the range of values in a specific columns of
a data table.
}
\examples{
# Create a check_range rule
rule1 <- DTAtools::DTARuleColRange(
  id = "rule1",
  columns = "age",
  range = c(18, 65)
)
}
