% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARuleColUnique-class.R
\name{DTARuleColUnique}
\alias{DTARuleColUnique}
\title{DTARuleColUnique Class}
\usage{
DTARuleColUnique(id, type, columns = NULL, description = NULL)
}
\arguments{
\item{id}{Character. A unique identifier for the rule.}

\item{columns}{columns that will be collectively checked if the
combinations are unique throughout the table}
}
\value{
An object of class \code{DTARule}.
}
\description{
Represents a single rule for validating data tables. The rule can be of various types,
such as \code{check_range}, \code{check_unique}, or \code{check_col_condition}
}
\examples{
# Create a check_unique rule
rule2 <- DTAtools::DTARuleColUnique(
  id = "rule2",
  columns = "id"
)
}
