% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTARules-class.R
\name{DTARules}
\alias{DTARules}
\title{DTARules Class}
\usage{
DTARules(rules)
}
\arguments{
\item{rules}{A list of DTARule objects.}
}
\value{
An object of class \code{DTARules}.
}
\description{
The \code{DTARules} class represents a collection of rules for validating data tables.
It provides functionality to define, store, and validate rules using a structured schema.
This class is designed to be used in data validation workflows, particularly for ensuring
compliance with predefined rules.
}
\examples{
# Define some example rules
<<<<<<< HEAD
example_rules <- list(
  DTARule(
    id = "rule1",
    type = "check_range",
    column = "age",
    range = c(18, 65)
  ),
  DTARule(
    id = "rule2",
    type = "check_unique",
    column = "id"
  )
)

=======
#example_rules <- list(
#    list(
#     id = "rule1",
#    type = "check_range",
#   column = "age",
#    range = c(18, 65)
#   ),
#  list(
#   id = "rule2",
#   type = "check_unique",
#   column = "id"
# )
# )
#
>>>>>>> f878237ed9fd646bc58a929fd9f17739761d5b68
# Create a DTARules object
#rules_obj <- DTARules(
#   rules = example_rules
#)

# Print the object
# print(rules_obj)
}
