% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFunctions.R
\name{export_specs_table}
\alias{export_specs_table}
\title{Generate table containing all DTA column specs}
\usage{
export_specs_table(
  DTAColumnSpecCollection,
  file = "dta_spec_table.docx",
  overwrite = FALSE,
  colnames = c("Variable Name", "Variable Label", "Type", "Length", "Format", "Nullable",
    "Description")
)
}
\arguments{
\item{DTAColumnSpecCollection}{DTAColumnSpecCollection object containing column specifications.}

\item{file}{Character. The name of the output Word file in which the DTA Spec Table shall be written. Default is "dta_spec_table.docx".}

\item{overwrite}{Logical. whether to overwrite the file.}

\item{colnames}{Vector. Vector containing column names of the DTA table. Default: c("Variable Name", "Variable Label", "Type", "Length", "Format", "Nullable", "Description")}
}
\value{
Flextable object that is saved.
}
\description{
This function takes a DTAColumnSpecCollection and generates a Word document containing a table with the DTA columns specs.
}
\examples{
\dontrun{
column_collection <- import_specs_from_yaml("spec.yaml")
export_specs_table(column_collection, "dta_spec_table.docx")
}
}
