% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpec-class.R
\name{get_arrow_schema_type}
\alias{get_arrow_schema_type}
\title{Get Arrow Schema Type}
\usage{
get_arrow_schema_type(x)
}
\arguments{
\item{x}{A DTAColumnSpec object.}
}
\value{
A character string representing the Arrow schema type.
}
\description{
Returns the corresponding Arrow schema type for a given DTAColumnSpec
object based on its \code{type} property.
}
\examples{
col <- DTAColumnSpec(id = "AGE", type = "Char")
get_arrow_schema_type(col)
}
