% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{get_column_ids-DTAColumnSpecCollection}
\alias{get_column_ids-DTAColumnSpecCollection}
\alias{get_column_ids}
\title{Get Names Method}
\usage{
get_column_ids(x, ...)
}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}

\item{...}{void}
}
\value{
A character vector of column names.
}
\description{
Method to get the names of columns in the collection.
}
\examples{
\dontrun{
names <- get_column_ids(collection)
}
}
