% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{get_max_number_of_files}
\alias{get_max_number_of_files}
\title{get max number of files}
\arguments{
\item{x}{An object of class DTAContainer}
}
\value{
number of files
}
\description{
Returns the number of files specified in the DTAFileInfo object.
}
\examples{
\dontrun{
column_format <- number_of_files(dtafileinfo)
}
}
