% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{get_metadata-DTAColumnSpecCollection}
\alias{get_metadata-DTAColumnSpecCollection}
\title{Get Metadata}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}
}
\value{
A list with metadata information
}
\description{
Method to get Metadata from DTAColumnSpecCollection
}
\examples{
\dontrun{
 get_metadata(collection)
}
}
