% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{get_table-DTADataSetTabular}
\alias{get_table-DTADataSetTabular}
\alias{get_table}
\title{Get table from DTADataSetTabular Object}
\usage{
get_table(x, id = 1, ...)
}
\arguments{
\item{x}{An object of class DTADataSet.}

\item{id}{Character or numeric. Name or index of the table to retrieve.}
}
\value{
An Arrow Table object.
}
\description{
Extract a table from the tables in a DTADataSetTabular object.
}
\examples{
ds <- create_example_DTADataSetTabular(2)
get_table(ds, 1)
get_table(ds, "tab1")
}
