% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAMetaData-helpers.R
\name{get_transmission_dates}
\alias{get_transmission_dates}
\title{Get Transmission Start and End Dates}
\usage{
get_transmission_dates(x)
}
\arguments{
\item{x}{A DTAMetaData object}
}
\value{
A list with elements first_transfer and last_transfer
}
\description{
Extract transmission transfer dates, whether they are Date objects or character phrases
}
\examples{
library(DTAtools)
md <- create_example_DTAMetaData(2)
get_transmission_dates(md)

}
