% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{getSpecs-DTAContainer}
\alias{getSpecs-DTAContainer}
\alias{getSpecs}
\title{Get DTAColumnSpecCollection (specs) from DTAContainer Object}
\usage{
getSpecs(x, ...)
}
\arguments{
\item{x}{An object of class DTAContainer.}
}
\value{
A DTAColumnSpecCollection object.
}
\description{
Method to extract the full DTAColumnSpecCollection from a DTAContainer object.
}
\examples{
\dontrun {
  getSpecs(container)
}
}
