% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{importDTAColumnSpecCollectionFromJson}
\alias{importDTAColumnSpecCollectionFromJson}
\title{Create DTAColumnSpecCollection from JSON File}
\usage{
importDTAColumnSpecCollectionFromJson(file)
}
\arguments{
\item{file}{Character. Path to the JSON file containing specifications.}
}
\value{
A DTAColumnSpecCollection object.
}
\description{
Parses a JSON file to extract column specifications and create a DTAColumnSpecCollection object.
}
