% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{labels-DTAContainer}
\alias{labels-DTAContainer}
\alias{labels}
\title{List of data labels within DTAContainer Object}
\usage{
labels(x, ...)
}
\arguments{
\item{x}{An object of class DTAContainer}

\item{...}{void}
}
\value{
A vector
}
\description{
Method to get a all data labels within a DTAContainer Object.
}
\examples{
\dontrun{
labels <- labels(dtadata)
}
}
