% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{labels-DTADataSetTabular}
\alias{labels-DTADataSetTabular}
\alias{labels}
\title{List of tables labels within DTADataSetTabular Object}
\usage{
labels(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}

\item{...}{void}
}
\value{
A vector
}
\description{
Method to get a all tables labels within a DTADataSetTabular Object.
}
\examples{
\dontrun{
labels <- labels(dtadata)
}
}
