% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{labels}
\alias{labels}
\title{List of tables labels within DTADataSetTabular Object}
\arguments{
\item{object}{An object of class DTADataSetTabular}

\item{...}{Additional arguments (not used).}
}
\value{
A character vector with table names.
}
\description{
Method to get a all tables labels within a DTADataSetTabular Object.
}
\examples{
ds <- create_example_DTADataSetTabular(2)
labels(ds)
}
