% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{load_file,DTAtools::DTADataSetTabular-method}
\alias{load_file,DTAtools::DTADataSetTabular-method}
\title{loads file}
\usage{
## S7 method for class <DTAtools::DTADataSetTabular>
load_file(x, file, index, name = basename(file))
}
\arguments{
\item{x}{An object of class DTADataSet}

\item{file}{file to be loaded}

\item{index}{of the filehandler in the files list}

\item{name}{file name, base name per default. is used to store the table under this name}
}
\value{
object of class DTADataSet with loaded data
}
\description{
Load the content of the file into dataset
}
\examples{
\dontrun{
column_format <- min_number_of_files(dtafiles)
}
}
