% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R, R/DTADataSet-class.R,
%   R/DTADataSetTabular-class.R
\name{load_file}
\alias{load_file}
\alias{load_file,DTAtools::DTADataSetTabular-method}
\title{Load a file into a DTA or DTADataSet object}
\usage{
load_file(x, ...)

## S7 method for class <DTAtools::DTA>
load_file(
  x,
  dataset,
  file,
  handler_index = 1,
  name = tools::file_path_sans_ext(basename(file)),
  ...
)

## S7 method for class <DTAtools::DTADataSet>
load_file(
  x,
  handler_index = 1,
  file,
  name = tools::file_path_sans_ext(basename(file)),
  ...
)

## S7 method for class <DTAtools::DTADataSetTabular>
load_file(
  x,
  file,
  handler_index,
  name = tools::file_path_sans_ext(basename(file))
)
}
\arguments{
\item{x}{An object of class DTADataSet}

\item{...}{Additional arguments passed through.}

\item{dataset}{Single character dataset name or numeric dataset index.}

\item{file}{file to be loaded}

\item{handler_index}{of the filehandler in the files list}

\item{name}{file name, base name per default. is used to store the table under this name}
}
\value{
The updated object.

The updated \code{DTA} object.

The updated dataset object.

object of class DTADataSet with loaded data
}
\description{
S7 generic. Dispatches to the appropriate method based on the class of \code{x}.
Use the \code{DTA} method to load a file into a named dataset within a full
DTA object; use the \code{DTADataSetTabular} method to load directly into a
standalone dataset.

Reads a file into one dataset contained in a \code{DTA} object by dataset
name or index.

Convenience wrapper that dispatches to \code{load_file()} for a dataset.

Load the content of the file into dataset
}
\examples{
file_handler <- DTAFileCSV(filename = "clinical_data.csv")
ds <- DTADataSetTabular(
  name = "demo",
  specs = create_example_DTAColumnSpecCollection(1),
  files = list(file_handler)
)
file <- system.file("extdata", "clinical_data.csv", package = "DTAtools")
ds <- DTAtools:::load_file(ds, file = file, handler_index = 1)
names(tables(ds))
}
