% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{matches_filename}
\alias{matches_filename}
\title{Matches Filename}
\arguments{
\item{x}{A \code{DTAFileInfo} object.}

\item{file}{A character string representing the name of the file to check against
the stored filename or pattern}
}
\value{
A logical value indicating whether the filename matches.
}
\description{
Checks if a given filename matches the pattern in a \code{DTAFileInfo} object.
}
\section{Methods}{

\describe{
\item{\code{DTAFileInfo}}{Returns \code{TRUE} if the filename matches the pattern.}
}
}

\examples{
\dontrun{
  file_info <- DTAFileInfo("file.txt")
  matches_filename(file_info, "file.txt")
}

}
