% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{matches_filename}
\alias{matches_filename}
\title{Check if Filename Matches Pattern}
\usage{
matches_filename(x, ...)
}
\arguments{
\item{x}{A DTAFileInfo object}

\item{file}{A character string representing the name of the file to check.}
}
\value{
Logical: TRUE if match, FALSE otherwise
}
\description{
Determines whether the provided filename matches the specified pattern.
}
\examples{
\dontrun{
matches_filename(dtafileinfo, "exact_file_name.tsv")
matches_filename(dtafileinfo, "\\\\.tsv$")
}
}
