% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{max_number_of_files-DTADataSet}
\alias{max_number_of_files-DTADataSet}
\title{get max number of files}
\arguments{
\item{x}{An object of class DTADataSet}
}
\value{
numeric: number of files
}
\description{
Returns the sum of max number of files specified all associated DTAFileInfo
objects.
}
\examples{
\dontrun{
column_format <- max_number_of_files(dtafileinfo)
}
}
