% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFile-class.R
\name{max_number_of_files}
\alias{max_number_of_files}
\title{Get max number of files}
\usage{
## S7 method for class <DTAtools::DTAFile>
max_number_of_files(x)
}
\arguments{
\item{x}{An object of class \code{DTAFile}.}
}
\value{
The number of files.
}
\description{
Returns the max number of files specified in a \code{DTAFile} object.
}
\section{Methods}{

\describe{
\item{\code{DTAFile}}{Returns the \code{max_number_of_files} property.}
}
}

\examples{
  file_info <- DTAFile("file.txt", number_of_files = 1)
  max_number_of_files(file_info)

}
