% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R
\name{metadata}
\alias{metadata}
\title{Get Metadata}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{An object of class DTA}
}
\value{
A list with metadata information
}
\description{
Method to get Metadata from a DTA object.
}
\examples{
library(DTAtools)
dta_obj <- create_example_DTA()
metadata(dta_obj)
}
