% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{min_number_of_files-DTAContainer}
\alias{min_number_of_files-DTAContainer}
\title{get min number of files}
\arguments{
\item{x}{An object of class DTAContainer}
}
\value{
numeric: number of files
}
\description{
Returns the sum of min number of files specified all associated DTAFileInfo
objects.
}
\examples{
\dontrun{
column_format <- number_of_files(dtafileinfo)
}
}
