% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFile-class.R
\name{min_number_of_files}
\alias{min_number_of_files}
\title{Get min number of files}
\arguments{
\item{x}{An object of class \code{DTAFile}.}
}
\value{
The number of files.
}
\description{
Returns the min number of files specified in a \code{DTAFile} object.
}
\section{Methods}{

\describe{
\item{\code{DTAFile}}{Returns the \code{min_number_of_files} property.}
}
}

\examples{
\dontrun{
  file_info <- DTAFile("file.txt", number_of_files = 1)
  min_number_of_files(file_info)
}

}
