% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{names}
\alias{names}
\title{Get Names Method}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}

\item{...}{Additional arguments (not used).}
}
\value{
A character vector of column names.
}
\description{
Method to get the names of columns in the collection.
}
\examples{
\dontrun{
names <- specids(collection)
}
}
