% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{number_of_files}
\alias{number_of_files}
\title{Generic: number_of_files}
\usage{
number_of_files(x, ...)
}
\arguments{
\item{x}{An object of class DTAFileInfo.}
}
\value{
Depends on the method.

Integer: number of files.
}
\description{
Generic function to get the number of files from an object.

Returns the number of files specified in the DTAFileInfo object.
}
\examples{
\dontrun{
number_of_files(dtafileinfo)
}
}
