% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R,
%   R/DTARule-class.R
\name{print}
\alias{print}
\title{print}
\usage{
print(x, ...)

print(x, ...)
}
\arguments{
\item{x}{An object of class DTARule}
}
\description{
Print overview for DTAColumnSpecCollection

Print overview for DTARule
}
\examples{
\dontrun{
 print(specs)
}
\dontrun{
 print(rule)
}
}
