% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA-class.R, R/DTAColumnSpec-class.R,
%   R/DTAColumnSpecCollection-class.R, R/DTAColumnSpecStructure-class.R,
%   R/DTAColumnSpecStructureSAS-class.R, R/DTADataSet-class.R,
%   R/DTAFile-class.R, R/DTAFileAny-class.R, R/DTADataSetTabular-class.R,
%   R/DTAFileCSV-class.R, R/DTAFileDelim-class.R, R/DTAFileTSV-class.R,
%   R/DTAMetaData-class.R, R/DTARule-class.R, R/DTARuleColCondition-class.R,
%   R/DTARuleColRange-class.R, R/DTARuleColUnique-class.R,
%   R/DTARuleGroupCondition-class.R
\name{print}
\alias{print}
\title{Print DTA Object}
\usage{
## S7 method for class <DTAtools::DTA>
print(x, ...)

## S7 method for class <DTAtools::DTAColumnSpec>
print(x, ...)

## S7 method for class <DTAtools::DTAColumnSpecCollection>
print(x, ...)

## S7 method for class <DTAtools::DTADataSet>
print(x, ...)

## S7 method for class <DTAtools::DTAFile>
print(x, ...)

## S7 method for class <DTAtools::DTAFileAny>
print(x, ...)

## S7 method for class <DTAtools::DTADataSetTabular>
print(x, ...)

## S7 method for class <DTAtools::DTAFileCSV>
print(x, ...)

## S7 method for class <DTAtools::DTAFileDelim>
print(x, ...)

## S7 method for class <DTAtools::DTAFileTSV>
print(x, ...)

## S7 method for class <DTAtools::DTAMetaData>
print(x, ...)

## S7 method for class <DTAtools::DTARule>
print(x, ...)

## S7 method for class <DTAtools::DTARuleColCondition>
print(x, ...)

## S7 method for class <DTAtools::DTARuleColRange>
print(x, ...)

## S7 method for class <DTAtools::DTARuleColUnique>
print(x, ...)

## S7 method for class <DTAtools::DTARuleGroupCondition>
print(x, ...)
}
\arguments{
\item{x}{An object of class DTARuleGroupCondition}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object

Invisibly returns the input object

Invisibly returns the input object

Invisibly returns the input object

Invisibly returns the input object

Invisibly returns the input object

Invisibly returns the input object
}
\description{
Print method for DTA objects.

S7 print method for DTAColumnSpec objects.

Print overview for DTAColumnSpecCollection

prints info of the column spec structure

prints info of the column spec structure

Print a summary of a DTADataSet object.

Print method for DTAFile objects.

Print method for DTAFileAny objects.

Print a summary of a DTADataSetTabular object.

Print method for DTAFileCSV objects.

Print method for DTAFileDelim objects.

Print method for DTAFileTSV objects.

Print method for DTAMetadata objects.

Print overview for DTARule

Print overview for DTADTARuleColConditionRule

Print overview for DTARuleColRange

Print overview for DTARuleColUnique

Print overview for DTARuleGroupCondition.
}
\examples{
dta_obj <- create_example_DTA()
print(dta_obj)
x <- create_example_DTAColumnSpecCollection()
print(x)
library(DTAtools)
print(create_example_DTADataSetTabular())
# do not use this, use derived classes instead, e.g.
# DTAFileCSV or DTAFileTSV
print(DTAFileCSV("example.csv"))
library(DTAtools)
print(DTAFileAny(filename = "study_report.pdf", extensions = "pdf"))

library(DTAtools)
print(create_example_DTADataSetTabular())
library(DTAtools)
print(create_example_DTAFileCSV())

library(DTAtools)
print(DTAFileDelim("example.tsv"))

library(DTAtools)
print(create_example_DTAFileTSV())

library(DTAtools)
print(create_example_DTAMetaData())

rule <- DTARuleFactory("rule1", "col_unique", columns = "SUBJID")
print(rule)
rule <- create_example_DTARuleColCondition()
print(rule)
rule <- create_example_DTARuleColRange()
print(rule)
rule <- create_example_DTARuleColUnique()
print(rule)
}
