% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfo-class.R
\name{read_file-DTAFileInfo}
\alias{read_file-DTAFileInfo}
\title{Read File for DTAFileInfo Objects}
\arguments{
\item{x}{A \code{DTAFileInfo} object containing file reading parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
An arrow table containing the contents of the file if the filename
matches; otherwise, returns \code{NULL}.
}
\description{
This is a protype funnction for reading in files.
}
