% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfoCSV-class.R
\name{read_file-DTAFileInfoCSV}
\alias{read_file-DTAFileInfoCSV}
\title{DTAFileInfoCSV Class Constructor}
\arguments{
\item{filename}{Character. Path to the CSV file.}

\item{number_of_files}{Integer. Number of files to be processed.
Defaults to 1.}

\item{sep}{Character. Field separator used in the CSV file.
Defaults to tab ("\\t").}

\item{has_header}{Logical. Indicates if the CSV file contains
a header row. Defaults to \code{TRUE}.}

\item{quote}{Character. Quoting character used in the CSV file.
Defaults to double quote ('"').}

\item{col_types}{Optional. Column types specification. Defaults
to \code{NULL}.}
}
\value{
An object of class \code{DTAFileInfoCSV}.
}
\description{
Defines the S7 class \code{DTAFileInfoCSV}, which extends \code{DTAFileInfo}
to represent metadata and configuration for CSV (Tab-Separated Values)
data files.
}
\seealso{
\code{\link{DTAFileInfo}}
}
