% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileInfoTSV-class.R
\name{read_file-DTAFileInfoTSV}
\alias{read_file-DTAFileInfoTSV}
\title{Read File for DTAFileInfoTSV Objects}
\arguments{
\item{x}{A \code{DTAFileInfoTSV} object containing file reading parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
A tibble containing the contents of the file if the filename
matches; otherwise, returns \code{NULL}.
}
\description{
Reads a TSV file using the parameters specified in a
\code{DTAFileInfoTSV} object. This method uses \code{readr::read_tsv}
for efficient TSV parsing.
}
\seealso{
\code{\link{readr::read_tsv}}
}
