% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFile-class.R
\name{read_file}
\alias{read_file}
\title{Read a file based on DTAFile}
\arguments{
\item{x}{A \code{DTAFile} object (or subclass) containing file reading
parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
An Arrow Table containing the file's contents.
}
\description{
Reads a data file using the parameters specified in a
\code{DTAFile} object or one of its subclasses.
}
\section{Methods}{

\describe{
\item{\code{DTAFile}}{This is a base implementation that throws an error,
as it must be implemented by a subclass.}
}
}

