% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAFileCSV-class.R
\name{read_file_execution-DTAFileCSV}
\alias{read_file_execution-DTAFileCSV}
\title{Read File for DTAFileCSV Objects}
\arguments{
\item{x}{A \code{DTAFileCSV} object containing file reading parameters.}

\item{file}{A character string specifying the path to the file to be read.}
}
\value{
A tibble containing the contents of the file if the filename
matches; otherwise, returns \code{NULL}.
}
\description{
Reads a CSV file using the parameters specified in a
\code{DTAFileCSV} object. This method uses \code{arrow::read_csv_arrow}
for efficient CSV parsing.
}
