% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateSchemaRules.R
\name{rule_check_range}
\alias{rule_check_range}
\title{Rule: check_range}
\usage{
rule_check_range(rule, df)
}
\arguments{
\item{rule}{DTA rule}

\item{df}{data}
}
\description{
Ensures a column's values fall within a specified numeric range.
}
