% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateSchemaRules.R
\name{rule_check_unique}
\alias{rule_check_unique}
\title{Rule: check_unique}
\usage{
rule_check_unique(rule, df)
}
\arguments{
\item{rule}{DTA rule}

\item{df}{data}
}
\description{
Ensures that all values in a column are unique.
}
