% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{rule_preview}
\alias{rule_preview}
\title{Preview Rules in a DTAColumnSpecCollection}
\usage{
rule_preview(x, ...)
}
\arguments{
\item{x}{A \code{DTAColumnSpecCollection} object.}
}
\value{
A character string representing a preview of the rules
}
\description{
Provides a preview of the rules contained within a \code{DTAColumnSpecCollection} object.
If the collection contains more than 5 rules, it returns the IDs of the first four rules,
an ellipsis, and the rules ID. If there are 5 or fewer rules, it returns all rules.
}
\examples{
library(DTAtools)
x <- create_example_DTAColumnSpecCollection()
rule_preview(x)
}
\seealso{
\code{\link{DTAColumnSpecCollection}}
}
