% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{rules-DTADataSet}
\alias{rules-DTADataSet}
\alias{rules}
\title{Get Rules}
\usage{
rules(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSet}
}
\value{
A list with rules information
}
\description{
Method to get Rules from DTADataSet.
}
\examples{
\dontrun{
rules(DTADataSet)
}
}
