% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{rules-DTADataSetTabular}
\alias{rules-DTADataSetTabular}
\alias{rules}
\title{Get Rules}
\usage{
rules(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}
}
\value{
A list with rules information
}
\description{
Method to get Rules from DTADataSet.
}
\examples{
\dontrun{
rules(DTADataSetTabular)
}
}
