% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{rules-DTADataSetTabular}
\alias{rules-DTADataSetTabular}
\title{Get Rules}
\usage{
## S7 method for class <DTAtools::DTADataSetTabular>
rules(x)
}
\arguments{
\item{x}{An object of class DTADataSetTabular}
}
\value{
A list with rules information
}
\description{
Method to get Rules from DTADataSet.
}
\examples{
ds <- create_example_DTADataSetTabular(2)
rules(ds)
}
