% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{rules}
\alias{rules}
\title{Get Rules}
\usage{
rules(x, ...)
}
\arguments{
\item{x}{An object of class DTAColumnSpecCollection.}
}
\value{
A DTARuleCollection object, or NULL if no rules are defined.
}
\description{
Method to get Rules from DTAColumnSpecCollection
}
\examples{
collection <- create_example_DTAColumnSpecCollection()
rules(collection)
}
