% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSet-class.R
\name{specs-DTADataSet}
\alias{specs-DTADataSet}
\alias{specs}
\title{Get DTAColumnSpecCollection (specs) from DTADataSet Object}
\usage{
specs(x, ...)
}
\arguments{
\item{x}{An object of class DTADataSet.}

\item{...}{void}
}
\value{
A DTAColumnSpecCollection object.
}
\description{
Method to extract the full DTAColumnSpecCollection from a DTADataSet object.
}
\examples{
\dontrun{
  specs(container)
}
}
