% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{specs_to_jsonschema}
\alias{specs_to_jsonschema}
\title{Convert DTAColumnSpec s to JSON Schema}
\usage{
specs_to_jsonschema(specs)
}
\arguments{
\item{columns}{A list containing column spec information}
}
\value{
A list representing the JSON Schema.
}
\description{
Converts a DTAColumnSpec s into a JSON Schema.
}
