% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{to_json_schema}
\alias{to_json_schema}
\title{Convert DTAColumnSpec s to JSON Schema}
\arguments{
\item{columns}{Column spec information}
}
\value{
A list representing the JSON Schema.
}
\description{
Converts a DTAColumnSpec s into a JSON Schema.
}
\examples{
library(DTAtools)
specs <- create_example_DTAColumnSpecCollection()
to_json_schema(specs)
}
