% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateSchemaRules.R
\name{validateRules}
\alias{validateRules}
\title{Validate Rules defined in DTAColumnSpecCollection and a Table}
\usage{
validateRules(DTAColumnSpecCollection, table)
}
\arguments{
\item{DTAColumnSpecCollection}{A DTAColumnSpecCollection object with rules defined.}

\item{table}{A table that will be validated}
}
\value{
a list containing the results of the rules validation
}
\description{
Validates the structure of table using predefined rules.
}
