% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTADataSetTabular-class.R
\name{validation_errors}
\alias{validation_errors}
\title{Retrieve Validation Errors for One Table}
\usage{
validation_errors(x, ...)
}
\arguments{
\item{x}{A \'DTADataSetTabular\' object.}

\item{table}{Character or numeric table identifier.}

\item{source}{Character. One of \'auto\', \'memory\', or \'artifact\'.}
}
\value{
A list with detailed validation output.
}
\description{
Returns detailed validation output for one table, either from in-memory
store or from persisted artifact.
}
