% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAColumnSpecCollection-class.R
\name{write_columns_to_json}
\alias{write_columns_to_json}
\title{Write DTAColumnSpecCollection to JSON File}
\usage{
write_columns_to_json(columns, file, pretty = TRUE)
}
\arguments{
\item{columns}{An object of class DTAColumnSpecCollection.}

\item{file}{Character. Path to the JSON file to write the specifications to.}

\item{pretty}{Logical. Whether to pretty-print the JSON. Default is TRUE.}
}
\value{
NULL. The function writes the DTAColumnSpecCollection to a JSON file.
}
\description{
This function writes a DTAColumnSpecCollection object to a JSON file.
}
\examples{
columns <- create_example_DTAColumnSpecCollection()
out_file <- tempfile(fileext = ".json")
write_columns_to_json(columns, out_file)
unlink(out_file)
}
