% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDocuments.R
\name{write_file_specification}
\alias{write_file_specification}
\title{Alias: Export File Dataset Specifications}
\usage{
write_file_specification(x, file, ...)
}
\description{
Convenience alias for write_dataset_metadata()
}
