% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFunctions.R
\name{write_metadata}
\alias{write_metadata}
\title{Write Metadata to File}
\usage{
write_metadata(file, table, write_to_file)
}
\arguments{
\item{file}{Path to the input data file.}

\item{table}{A data.frame that was written to the file.}

\item{write_to_file}{Logical. Whether to write information to an additional file.}
}
\value{
Invisibly returns TRUE if successful.
}
\description{
Writes metadata (MD5 checksum, number of rows and columns) to a separate file.
}
