% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTAContainer-class.R
\name{write_table_to_file}
\alias{write_table_to_file}
\title{Write DTA Table to File}
\usage{
write_table_to_file(
  DTAContainer,
  table,
  filename,
  arrange_by = "all",
  arrange_desc = FALSE,
  sep = "\\t",
  na = "",
  row.names = FALSE,
  overwrite = TRUE,
  quote = FALSE,
  compression = c("none", "gzip"),
  get_md5sum = TRUE,
  write_md5sum_to_file = TRUE,
  ...
)
}
\arguments{
\item{DTAContainer}{An object of class DTAContainer.}

\item{table}{Character. The name of the table within the DTAContainer object to write.}

\item{filename}{Character. The name of the file to write to.}

\item{arrange_by}{Character vector. Columns to arrange the table by. NULL, Table won't be arranged. "all" (Default) -> Table will be arranged by all columns.}

\item{arrange_desc}{Logical. Whether to arrange the table in descending order. Default is FALSE.}

\item{sep}{Character. The field separator string. Default is ",".}

\item{na}{Character. The string to use for missing values in the data. Default is "".}

\item{row.names}{Logical. Row names provided.}

\item{overwrite}{Logical. Whether to overwrite the file if it exists. Default is FALSE.}

\item{quote}{Logical. Use of quotes}

\item{compression}{Character. Compression method, either "none" or "gzip". Default is "none".}

\item{get_md5sum}{Logical. Whether to calculate and print the MD5 checksum of the file. MD5SUM and number of rows and columns of file will be also saved in an additional file. Default is TRUE.}

\item{write_md5sum_to_file}{Logical. Whether to calculate and print the MD5 checksum of the file. MD5SUM and number of rows and columns of file will be also saved in an additional file. Default is TRUE.}

\item{...}{Additional arguments passed to write.table.}
}
\value{
NULL. The function writes the table to a file.
}
\description{
Write a named DTA table saved in a DTAContainer object to a file.
}
\examples{
\dontrun{
write_table_to_file(dtadata, table = "my_table", filename = "table.tsv.gz",
                 sep = "\t", arrange_by = c("STUDYID", "VISIT"))
}
}
